% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someviz.R
\name{add_tx}
\alias{add_tx}
\title{Use geom_point to enhance a visualization with viewShape}
\usage{
add_tx(sdobj, featurename, size = 0.1)
}
\arguments{
\item{sdobj}{SpatialData instance}

\item{featurename}{character(1) name of a shapes component of sdobj}

\item{size}{numeric(1) target size for glyph}
}
\description{
Use geom_point to enhance a visualization with viewShape
}
\examples{
example(use_sdio) # produces br2fov
viewShape(br2fov, "cell_boundaries") + add_tx(sdobj, "EPCAM")
}
