% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateImage.R
\name{aggregateImage}
\alias{aggregateImage}
\title{Aggregate to `SingleCellExperiment`}
\usage{
aggregateImage(x, which = 1, fun = mean)
}
\arguments{
\item{fun}{Function to use for aggregation.}

\item{data}{A \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\description{
...
}
\examples{
library(ggplot2)
library(SingleCellExperiment)
path <- file.path("extdata", "mibitof")
path <- system.file(path, package = "SpatialData")
spd <- readSpatialData(path)
sce <- aggregateImage(spd)
cd <- data.frame(colData(sce), z = assay(sce)[1, ])
ggplot(cd, aes(x, y, col = z)) + geom_point() +
  scale_color_viridis_c() + scale_y_reverse()

}
