% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andalusia.R
\docType{data}
\name{andalusia_predictors}
\alias{andalusia_predictors}
\title{Predictor variable names for andalusia dataset}
\format{
A character vector of length 20.
}
\usage{
data(andalusia_predictors)
}
\description{
Character vector of 20 predictor variable names from \link{andalusia},
covering Landsat reflectance (7), rainfall (2), solar radiation (2),
temperature (4), and topography (5).
}
\seealso{
Other andalusia: 
\code{\link{andalusia}},
\code{\link{andalusia_extra}()},
\code{\link{andalusia_responses}}
}
\concept{andalusia}
\keyword{datasets}
