% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andalusia.R
\docType{data}
\name{andalusia_responses}
\alias{andalusia_responses}
\title{Response variable names for andalusia dataset}
\format{
A character vector of length 2.
}
\usage{
data(andalusia_responses)
}
\description{
Character vector of length 2 containing the names of the
response variables in \link{andalusia}: \code{"species"} (character, species name or
\code{"background"}) and \code{"presence"} (binary integer, 1 = confirmed species presence).
}
\seealso{
Other andalusia: 
\code{\link{andalusia}},
\code{\link{andalusia_extra}()},
\code{\link{andalusia_predictors}}
}
\concept{andalusia}
\keyword{datasets}
