% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdArray.R
\name{Array-methods}
\alias{Array-methods}
\alias{data,ImageArray-method}
\alias{data,LabelArray-method}
\alias{dim,ImageArray-method}
\alias{dim,LabelArray-method}
\alias{length,ImageArray-method}
\alias{length,LabelArray-method}
\alias{data,sdArray-method}
\alias{dim,sdArray-method}
\alias{length,sdArray-method}
\alias{data_type,sdArray-method}
\alias{data_type,DelayedArray-method}
\title{Methods for `ImageArray` and `LabelArray` class}
\usage{
\S4method{data}{sdArray}(x, k = 1)

\S4method{dim}{sdArray}(x)

\S4method{length}{sdArray}(x)

\S4method{data_type}{sdArray}(x)

\S4method{data_type}{DelayedArray}(x)
}
\arguments{
\item{x}{\code{ImageArray} or  \code{LabelArray}}

\item{k}{scalar index specifying which scale to extract.}
}
\value{
\code{ImageArray}
}
\description{
Methods for `ImageArray` and `LabelArray` class
}
\examples{
library(SpatialData.data)
zs <- get_demo_SDdata("merfish")

# helper that gets path to first element in layer 'l' 
fn <- \(l) list.files(file.path(zs, l), full.names=TRUE)[1]
  
# read individual element  
(ia <- readImage(fn("images")))

}
