% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{available_10x_xen_zips}
\alias{available_10x_xen_zips}
\title{use 'paws::s3' to interrogate an NSF Open Storage Network 
bucket for zipped 10x-produced Xenium outputs}
\usage{
available_10x_xen_zips()
}
\description{
use 'paws::s3' to interrogate an NSF Open Storage Network 
bucket for zipped 10x-produced Xenium outputs
}
\examples{
if (requireNamespace("paws")) {
  available_10x_xen_zips()
}
}
