% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{Breast2fov_10x}
\alias{Breast2fov_10x}
\title{Retrieve 10x-trimmed breast cancer demonstration data, "two fields of view"}
\usage{
Breast2fov_10x(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve 10x-trimmed breast cancer demonstration data, "two fields of view"
}
\note{
From `https://www.10xgenomics.com/support/software/xenium-onboard-analysis/latest/resources/xenium-example-data`.
}
\examples{
br = Breast2fov_10x()
if (requireNamespace("SpatialData.plot")) {
 SpatialData.plot::plotSpatialData() + SpatialData.plot::plotShape(br)
}
}
