% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coercion}
\alias{coercion}
\title{`SpatialData` coercion}
\arguments{
\item{x}{A \code{\link{SpatialData}} object.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
it depends...
}
\description{
...
}
\examples{
# TODO
}
\author{
Helena L. Crowell
}
