% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{ColorectalCarcinomaMIBITOF}
\alias{ColorectalCarcinomaMIBITOF}
\title{Retrieve colorectal carcinoma MIBI-TOF experiment.}
\usage{
ColorectalCarcinomaMIBITOF(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve colorectal carcinoma MIBI-TOF experiment.
}
\note{
From spatialdata archive citing `Hartmann, F. J. et al. Single-cell metabolic profiling of human cytotoxic T cells. Nat. Biotechnol. (2020) doi:10.1038/s41587-020-0651-8.`
}
\examples{
ColorectalCarcinomaMIBITOF()
}
