% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{communities_extra_2010}
\alias{communities_extra_2010}
\title{Download Environmental Raster for communities - 2010}
\usage{
communities_extra_2010(dir = ".", quiet = TRUE)
}
\arguments{
\item{dir}{(optional, character) Directory to save the file. Defaults to
the current working directory.}

\item{quiet}{(optional, logical) If \code{TRUE} (default), suppresses
download messages.}
}
\value{
SpatRaster object.
}
\description{
Downloads and reads the baseline (2010) environmental raster associated with the \link{communities}
dataset from the
\href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository.
Requires the \pkg{terra} package.
}
\examples{
\dontrun{
communities_env_2010 <- communities_extra_2010()
communities_env_2010
}
}
\seealso{
Other communities: 
\code{\link{communities}},
\code{\link{communities_extra_2050}()},
\code{\link{communities_extra_2100}()},
\code{\link{communities_predictors}},
\code{\link{communities_responses}}
}
\concept{communities}
