% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{communities_extra_2100}
\alias{communities_extra_2100}
\title{Download Environmental Raster for communities - 2100}
\usage{
communities_extra_2100()
}
\value{
SpatRaster object.
}
\description{
Downloads the future climate (2100) raster associated with the \link{communities} dataset from the \href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository. Writes the file \code{communities_2100.tif} in the working directory and returns it as a \code{spatRaster} object.
}
\seealso{
Other communities:
\code{\link{communities}},
\code{\link[=communities_extra_2010]{communities_extra_2010()}},
\code{\link[=communities_extra_2050]{communities_extra_2050()}},
\code{\link{communities_predictors}},
\code{\link{communities_responses}}
}
\concept{communities}
