% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\docType{data}
\name{communities_predictors}
\alias{communities_predictors}
\title{Predictor variable names for the dataset \code{communities}}
\format{
A character vector of length 9.
}
\usage{
data(communities_predictors)
}
\description{
Character vector of 9 predictor variable names from \link{communities}.
}
\examples{
data(communities_predictors)
communities_predictors
}
\seealso{
Other communities: 
\code{\link{communities}},
\code{\link{communities_extra_2010}()},
\code{\link{communities_extra_2050}()},
\code{\link{communities_extra_2100}()},
\code{\link{communities_responses}}
}
\concept{communities}
\keyword{datasets}
