% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{coord-trans}
\alias{coord-trans}
\alias{ct_transform}
\alias{ct_translate}
\alias{ct_scale}
\alias{ct_rotate}
\title{Coordinate transformations}
\usage{
ct_transform(x, i, ...)

ct_translate(x, t, ...)

ct_scale(x, t, ...)

ct_rotate(x, t, ...)
}
\arguments{
\item{x}{\code{SpatialData} element.}

\item{i}{target coordinate system.}

\item{...}{ignored.}

\item{t}{transformation data.}
}
\value{
Object of same type as input.
}
\description{
Coordinate transformations
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)
gg <- sd_plot()

# point

x <- sd@points[[1]]
sd@points$a <- ct_translate(x, c(-10, 10))
sd@points$b <- ct_scale(x, c(0.5, 1.2))
sd@points$c <- ct_rotate(x, 5)

gg + 
  sd_plot_point(sd, 1, col=1) + 
  sd_plot_point(sd, "a", col=2) + 
  sd_plot_point(sd, "b", col=3) + 
  sd_plot_point(sd, "c", col=4) 

# shape

x <- sd@shapes[[2]]
sd@shapes$a <- ct_translate(x, c(-10, 10))
sd@shapes$b <- ct_scale(x, c(0.5, 1.2))
sd@shapes$c <- ct_rotate(x, 5)

gg + 
  sd_plot_shape(sd, 2, col=1) + 
  sd_plot_shape(sd, "a", col=2) + 
  sd_plot_shape(sd, "b", col=3) + 
  sd_plot_shape(sd, "c", col=4) 
}
