% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\alias{addCT}
\alias{rmvCT}
\alias{rmvCT,SpatialDataElement-method}
\alias{rmvCT,Zattrs-method}
\alias{addCT,SpatialDataElement-method}
\alias{addCT,Zattrs-method}
\title{Coordinate transformations}
\usage{
\S4method{rmvCT}{SpatialDataElement}(x, i)

\S4method{rmvCT}{Zattrs}(x, i)

\S4method{addCT}{SpatialDataElement}(x, name, type, data)

\S4method{addCT}{Zattrs}(x, name, type = "identity", data = NULL)
}
\arguments{
\item{x}{\code{\link{SpatialData}} element 
or \code{\link{Zattrs}} of an element}

\item{i}{scalar charcter or integer; 
name or index of coordinate space}

\item{name}{character(1); name of coordinate space}

\item{type}{character(1); type of transformation}

\item{data}{transformation data; size and shape depend on transformation and
element type (e.g., numeric(1) for rotation, numeric(2) for scaling in 2D)}
}
\description{
Coordinate transformations
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# view available coordinate transformations
CTdata(z <- meta(label(x)))

# add
addCT(z, "scale", "scale", c(12, 34)) # can't overwrite
CTdata(addCT(z, "new", "translation", c(12, 34)))

# rmv
CTdata(rmvCT(z, 2)) # by index
CTdata(rmvCT(z, "scale")) # by name
CTdata(rmvCT(z, 1)) # identity is protected
}
