% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{.coord2graph}
\alias{.coord2graph}
\title{CS graph representation}
\usage{
.coord2graph(x)
}
\arguments{
\item{x}{\code{SpatialData} object}
}
\description{
CS graph representation
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# element-wise
g <- SpatialData:::.coord2graph(image(x))
SpatialData:::.get_path(g, "self", "global")

# object-wide
g <- SpatialData:::.coord2graph(x)
graphics.off(); graph::plot(g)
# retrieve transformation from element to target space
graph::edgeData(g, "blobs_labels", "translation", "data")

}
