% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\alias{coord}
\alias{coord<-}
\alias{coords<-}
\alias{coordNames}
\alias{coords<-,SpatialDataElement,data.frame-method}
\alias{coords,Zattrs-method}
\alias{coords,SpatialDataElement-method}
\alias{coords,list-method}
\alias{coords,SpatialData-method}
\alias{coord,Zattrs-method}
\alias{coord,SpatialDataElement-method}
\alias{coordNames,Zattrs-method}
\alias{coordNames,SpatialDataElement-method}
\alias{coordNames,list-method}
\alias{coordNames,SpatialData-method}
\alias{coords<-,Zattrs,ANY-method}
\title{coordinate systems}
\usage{
\S4method{coords}{SpatialDataElement,data.frame}(x) <- value

\S4method{coords}{Zattrs}(x)

\S4method{coords}{SpatialDataElement}(x)

\S4method{coords}{list}(x)

\S4method{coords}{SpatialData}(x)

\S4method{coord}{Zattrs}(x, i = 1)

\S4method{coord}{SpatialDataElement}(x, i = 1)

\S4method{coordNames}{Zattrs}(x)

\S4method{coordNames}{SpatialDataElement}(x)

\S4method{coordNames}{list}(x)

\S4method{coordNames}{SpatialData}(x)

\S4method{coords}{Zattrs,ANY}(x) <- value
}
\arguments{
\item{x}{A \code{\link{SpatialData}} element or \code{Zattrs} objects.}

\item{value}{Replacement value.}

\item{i}{Scalar integer or character string
specifying which element to extract/replace.}
}
\value{
If \code{x} is a \code{Zattrs} object,
\code{CoordNames} returns a character vector,
\code{coord(s)} returns a \code{data.frame}.
If \code{x} is a \code{SpatialData} object
or a list of \code{SpatialData} elements,
only shared coordinate systems are returned.
}
\description{
...
}
\examples{
dir <- file.path("extdata", "blobs")
dir <- system.file(dir, package="SpatialData")
spd <- readSpatialData(dir)

# all available
coordNames(image(spd)) 
coordNames(label(spd)) 

# shared only  
lys <- list(image(spd), label(spd))
coordNames(lys)
coords(lys)

}
\author{
Helena L. Crowell
}
