% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr_utils.R
\name{create_zarr}
\alias{create_zarr}
\title{create_zarr}
\usage{
create_zarr(store, version = 2)
}
\arguments{
\item{store}{The location of the Zarr store}

\item{version}{Zarr version}
}
\value{
`NULL`
}
\description{
Create Zarr store
}
\examples{
store <- tempfile(fileext = ".zarr")
create_zarr(store = store)
dir.exists(store)

}
