% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTgraph.R
\name{CTgraph}
\alias{CTgraph}
\alias{CTpath}
\alias{CTplot}
\alias{CTgraph,SpatialData-method}
\alias{CTgraph,SpatialDataElement-method}
\alias{CTgraph,ANY-method}
\alias{CTpath,SpatialData-method}
\alias{CTpath,SpatialDataElement-method}
\alias{CTpath,ANY-method}
\title{Coord. trans. graph}
\usage{
\S4method{CTgraph}{SpatialData}(x)

\S4method{CTgraph}{SpatialDataElement}(x)

\S4method{CTgraph}{ANY}(x)

\S4method{CTpath}{SpatialData}(x, i, j)

\S4method{CTpath}{SpatialDataElement}(x, j)

\S4method{CTpath}{ANY}(x)

CTplot(g, cex = 0.5, fac = 2, max = 10)
}
\arguments{
\item{x}{\code{SpatialData}, an element, or \code{Zattrs}.}

\item{i}{character string; name of source node.}

\item{j}{character string; name of target coordinate space.}

\item{g}{base R graph; extracted with \code{CTgraph}.}

\item{cex}{scalar numeric; controls fontsize of node labels.}

\item{fac, max}{scalar numeric; node labels with \code{nchar>max}
are split and hyphenated at position \code{floor(nchar/fac)}}
}
\description{
Coord. trans. graph
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# object-wide
g <- CTgraph(x)
CTplot(g)   

# one element
y <- label(x)
g <- CTgraph(y)
CTplot(g) 

# retrieve transformation(s) 
# from element to target space
CTpath(x, "blobs_labels", "sequence")
}
