% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\docType{data}
\name{demodata_df}
\alias{demodata_df}
\title{a data.frame with information about available resources}
\format{
An object of class \code{data.frame} with 11 rows and 5 columns.
}
\usage{
demodata_df
}
\description{
a data.frame with information about available resources
}
\note{
This information was scraped from [scverse spatialdata](https://spatialdata.scverse.org/en/latest/tutorials/notebooks/datasets/README.html) download site on 5 Dec 2024.
}
\keyword{datasets}
