% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_to_ext.R
\name{do_tx_to_ext}
\alias{do_tx_to_ext}
\title{Use Python's 'spatialdata' 'transform_to_data_extent' on a spatialdata zarr store}
\usage{
do_tx_to_ext(srcdir, dest, coordinate_system, ...)
}
\arguments{
\item{srcdir}{character(1) path to folder holding a zarr store}

\item{dest}{character(1) a path to a desired destination for new zarr representation}

\item{coordinate_system}{character(1) defaults to "global"}

\item{...}{arguments for passage to python `transform_to_data_extent`;
can include "maintain_positioning" (logical (1)) or numerics for
target_unit_to_pixels, target_width, target_height, target_depth.}
}
\description{
Use Python's 'spatialdata' 'transform_to_data_extent' on a spatialdata zarr store
}
\examples{
src <- system.file("extdata", "blobs.zarr", package="SpatialData")
td <- tempfile()
do_tx_to_ext(
  srcdir=src, dest=td, 
  coordinate_system="global",
  maintain_positioning=FALSE, 
  target_width=400.)
(sd <- readSpatialData(td))

}
