% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drylands_predictors}
\alias{drylands_predictors}
\title{Predictor variable names for drylands examples}
\format{
A character vector of length 60.
}
\usage{
data(drylands_predictors)
}
\description{
Character vector of 60 predictor variable names from \code{drylands}, covering
soil properties (10), soil water (4), topography (7), geography (1), vegetation (4),
atmosphere (8), climate/aridity (24), and ecoregion morphology (2).
}
\seealso{
Other data: 
\code{\link{drylands}},
\code{\link{drylands_responses}},
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}},
\code{\link{plants}},
\code{\link{plants_predictors}},
\code{\link{plants_response}},
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}}
}
\concept{data}
\keyword{datasets}
