% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drylands.R
\docType{data}
\name{drylands_responses}
\alias{drylands_responses}
\title{Response variable names for drylands examples}
\format{
A character vector of length 5.
}
\usage{
data(drylands_responses)
}
\description{
Character vector containing the names of the five response variables in \code{drylands}:
species, genuses, families, classes, and phyla counts per ecoregion.
}
\seealso{
Other data: 
\code{\link{drylands}},
\code{\link{drylands_predictors}},
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}},
\code{\link{plantae}},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}},
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}}
}
\concept{data}
\keyword{datasets}
