% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{extent}
\alias{extent}
\alias{extent,SpatialData-method}
\alias{extent,sdArray-method}
\alias{extent,sdFrame-method}
\title{Spatial element extent}
\usage{
\S4method{extent}{SpatialData}(x, i = 1)

\S4method{extent}{sdArray}(x, i = 1)

\S4method{extent}{sdFrame}(x, i = 1)
}
\arguments{
\item{x}{a \code{SpatialData} element (any but table).}

\item{i}{scalar integer or string; target coordinate space.}
}
\value{
Length-2 list with numeric x and y ranges.
}
\description{
Spatial element extent
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# object-wide
extent(x)

# element-wise
extent(image(x))
extent(point(x))
extent(shape(x))

# with transformation(s)
extent(label(x), "scale")
extent(label(x), "translation")
}
