% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{get_demo_SD}
\alias{get_demo_SD}
\title{this function consolidates the retrieval and caching and
transformation of scverse-curated Zarr archives and 10x-curated
Xenium archives}
\usage{
get_demo_SD(patt, cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{patt}{character(1) sufficient to identify a resource in OSN}

\item{cache}{like BiocFileCache}
}
\description{
this function consolidates the retrieval and caching and
transformation of scverse-curated Zarr archives and 10x-curated
Xenium archives
}
\examples{
(br2fov <- get_demo_SD("Breast_2fov"))
}
