% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{.get_path}
\alias{.get_path}
\title{get transformations path}
\usage{
.get_path(g, i, j)
}
\arguments{
\item{g}{\code{\link[graph]{graphAM}}}

\item{i, j}{source and target node label}
}
\description{
get transformations path
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)
g <- SpatialData:::.coord2graph(x)
SpatialData:::.get_path(g, "blobs_labels", "sequence")

}
