% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{getRegionData,SpatialData,character-method}
\alias{getRegionData,SpatialData,character-method}
\title{Get Region Data}
\usage{
\S4method{getRegionData}{SpatialData,character}(x, re)
}
\arguments{
\item{x}{A SpatialData object from which the region data is to be retrieved.}

\item{re}{A character string specifying the region for which data is to be retrieved.}
}
\value{
A SingleCellExperiment object containing the data for the specified region.
}
\description{
This method retrieves data for a specified region from a SpatialData object.
}
\examples{
# Assuming `spatial_data` is a SpatialData object and "region_name" is the name of the region
region_data <- getRegionData(spatial_data, "region_name")
}
