% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr_utils.R
\name{getZarrArrayPath}
\alias{getZarrArrayPath}
\title{getZarrArrayPath}
\usage{
getZarrArrayPath(a)
}
\arguments{
\item{a}{ZarrArray}
}
\description{
returns the path of a zarr store associated with ZarrArray
}
