% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{ImageArray}
\alias{ImageArray}
\alias{channels}
\alias{channels,ImageArray-method}
\alias{channels,ANY-method}
\alias{[,ImageArray,ANY,ANY,ANY-method}
\title{The `ImageArray` class}
\usage{
ImageArray(data = list(), meta = Zattrs(), metadata = list(), ...)

\S4method{channels}{ImageArray}(x, ...)

\S4method{channels}{ANY}(x, ...)

\S4method{[}{ImageArray,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)
}
\arguments{
\item{data}{list of \code{\link[ZarrArray]{ZarrArray}}s}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{ImageArray}}

\item{i, j}{indices specifying elements to extract.}

\item{k}{scalar index specifying which scale to extract.}

\item{drop}{ignored.}
}
\value{
\code{ImageArray}
}
\description{
The `ImageArray` class
}
\examples{
library(SpatialData.data)
zs <- get_demo_SDdata("merfish")
pa <- file.path(zs, "images", "rasterized")
(ia <- readImage(pa))

}
