% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction_extra}
\alias{interaction_extra}
\title{Download Environmental Raster for interaction}
\usage{
interaction_extra()
}
\value{
SpatRaster object.
}
\description{
Downloads and reads the environmental raster associated with the \link{interaction} dataset from the \href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository. Requires the \pkg{terra} package. Writes the file \code{sierra_nevada_env.tif} to the working directory and returns a \code{spatRaster} object.
}
\seealso{
Other interaction:
\code{\link{interaction}},
\code{\link{interaction_predictors}},
\code{\link{interaction_responses}}
}
\concept{interaction}
