% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\docType{data}
\name{interaction_predictors}
\alias{interaction_predictors}
\title{Predictor variable names for interaction dataset}
\format{
A character vector of length 10.
}
\usage{
data(interaction_predictors)
}
\description{
Character vector of 10 predictor variable names from \link{interaction},
covering remote sensing (3: Landsat NDVI and two PCA scores), climate (5: annual
rainfall, solar radiation, mean annual temperature, maximum summer temperature,
minimum winter temperature), and topography (2: topographic complexity and
topographic position).
}
\seealso{
Other interaction: 
\code{\link{interaction}()},
\code{\link{interaction_extra}()},
\code{\link{interaction_responses}}
}
\concept{interaction}
\keyword{datasets}
