% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{JanesickBreastVisiumEnh}
\alias{JanesickBreastVisiumEnh}
\title{Retrieve a version of Janesick FFPE breast cancer experiment, Visium platform}
\usage{
JanesickBreastVisiumEnh(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve a version of Janesick FFPE breast cancer experiment, Visium platform
}
\note{
From spatialdata archive citing the biorxiv antecedent to `https://www.nature.com/articles/s41467-023-43458-x`
This is annotated as Visium associated xenium, and has been enhanced
with cell type annotation
}
\examples{
JanesickBreastVisiumEnh()
}
