% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{JanesickBreastXeniumRep2}
\alias{JanesickBreastXeniumRep2}
\title{Retrieve replicate 2 of Janesick FFPE breast cancer experiment, Xenium platform}
\usage{
JanesickBreastXeniumRep2(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve replicate 2 of Janesick FFPE breast cancer experiment, Xenium platform
}
\note{
From spatialdata archive citing the biorxiv antecedent to `https://www.nature.com/articles/s41467-023-43458-x`
This has been enhanced
with cell type annotation.
}
\examples{
jj = JanesickBreastXeniumRep2()
jj
if (requireNamespace("SpatialData.plot")) {
  SpatialData.plot::plotSpatialData() + SpatialData.plot::plotShape(jj)
}
}
