% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelArray.R
\name{LabelArray}
\alias{LabelArray}
\alias{[,LabelArray,ANY,ANY,ANY-method}
\title{The \code{LabelArray} class}
\usage{
LabelArray(
  data = list(),
  meta = Zattrs(label = TRUE),
  version = image(sdFormat(0.1)),
  metadata = list(),
  scale_factors = NULL,
  ...
)

\S4method{[}{LabelArray,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{data}{list of \code{\link[ZarrArray]{ZarrArray}}s}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{LabelArray}}

\item{i, j}{indices specifying elements to extract.}

\item{drop}{ignored.}

\item{multiscale}{if TRUE (and \code{data} is not a list), 
multiscale image will be generated.}

\item{axes}{axes}
}
\value{
\code{LabelArray}
}
\description{
The \code{LabelArray} class stores \code{SpatialData} elements from its 
\code{"labels"} layers. These are represented as a \code{ZarrMatrix} 
(\code{data} slot) associated with .zattrs stored as \code{\link{Zattrs}} 
(\code{meta} slot); a list of \code{metadata} stores other arbitrary info.

Currently defined methods (here, \code{x} is a \code{LabelArray}):
\itemize{
\item \code{data/meta(x)} to access underlying \code{ZarrMatrix/Zattrs}
\item \code{dim(x)} returns the dimensions of \code{data(x)}
}
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- file.path(x, "labels", "blobs_labels")

(y <- readLabel(x))
y[1:10, 1:10]
meta(y)

}
