% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linaria.R
\docType{data}
\name{linaria_predictors}
\alias{linaria_predictors}
\title{Predictor variable names for the dataset \code{linaria}}
\format{
A character vector of length 20.
}
\usage{
data(linaria_predictors)
}
\description{
Character vector of 20 predictor variable names from \link{linaria},
covering Landsat reflectance (7), rainfall (2), solar radiation (2),
temperature (4), and topography (5).
}
\examples{
data(linaria_predictors)
linaria_predictors
}
\seealso{
Other linaria:
\code{\link{linaria}},
\code{\link[=linaria_extra]{linaria_extra()}},
\code{\link{linaria_responses}}
}
\concept{linaria}
\keyword{datasets}
