% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linaria.R
\docType{data}
\name{linaria_responses}
\alias{linaria_responses}
\title{Response variable names for linaria dataset}
\format{
A character vector of length 2.
}
\usage{
data(linaria_responses)
}
\description{
Character vector of length 2 containing the names of the
response variables in \link{linaria}: \code{"linaria_nigricans"} (binary integer,
1 = confirmed \emph{Linaria nigricans} presence) and \code{"greenhouses"} (binary
integer, 1 = greenhouse presence).
}
\seealso{
Other linaria: 
\code{\link{linaria}},
\code{\link{linaria_extra}()},
\code{\link{linaria_predictors}}
}
\concept{linaria}
\keyword{datasets}
