% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{LungAdenocarcinomaMCMICRO}
\alias{LungAdenocarcinomaMCMICRO}
\title{Retrieve small cell lung adenocarcinoma sample assayed with mcmicro}
\usage{
LungAdenocarcinomaMCMICRO(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve small cell lung adenocarcinoma sample assayed with mcmicro
}
\note{
From scverse spatialdata archive citing `https://www.nature.com/articles/s41592-021-01308-y`.
}
\examples{
LungAdenocarcinomaMCMICRO()
}
