% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{merfish_demo_path}
\alias{merfish_demo_path}
\title{check cache for merfish.zarr.zip and return path; 
retrieve/stash/return path if not found}
\usage{
merfish_demo_path(
  cache = BiocFileCache::BiocFileCache(),
  zipname = "merfish.zarr.zip"
)
}
\arguments{
\item{cache}{defaults to BiocFileCache::BiocFileCache(), 
will serve as destination or source for data}

\item{zipname}{character(1) name of zip archive to find}
}
\value{
Returns character(1) path to cached zip file.
}
\description{
check cache for merfish.zarr.zip and return path; 
retrieve/stash/return path if not found
}
