% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock}
\alias{mock}
\alias{mockImage}
\alias{mockLabel}
\alias{mockShape}
\alias{mockPoint}
\title{mock data}
\usage{
mockImage()

mockLabel(n = 10)

mockShape(n = 10, type = c("circle", "polygon"))

mockPoint(n = 100)
}
\arguments{
\item{n}{Scalar integer specifying the number of points/shapes to generate.}

\item{type}{Character string specifying which type of shapes to generate.}
}
\value{
A \code{SpatialData} element.
}
\description{
...
}
\examples{
x <- SpatialData(
  images=list(noise=mockImage()),
  labels=list(squares=mockLabel()),
  points=list(dots=mockPoint()),
  shapes=list(
    blobs=mockShape(type="circle"), 
    wedges=mockShape(type="polygon")))
plotImage(x)
plotLabel(x)
plotShape(x, i="blobs")
plotShape(x, i="wedges")
plotPoint(x, col="z")

}
\author{
Helena L. Crowell
}
