% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{MouseBrainMERFISH}
\alias{MouseBrainMERFISH}
\title{Retrieve small cell lung adenocarcinoma sample assayed with mcmicro}
\usage{
MouseBrainMERFISH(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve small cell lung adenocarcinoma sample assayed with mcmicro
}
\note{
From scverse spatialdata archive citing `Moffitt, J. R. et al. Molecular, spatial, and functional single-cell profiling of the hypothalamic preoptic region. Science 362, (2018).`
}
\examples{
if (requireNamespace("SpatialData.plot")) {
  library(SpatialData.plot)
  mb  = MouseBrainMERFISH()
  plotSpatialData() + plotImage(mb) + plotShape(mb)
}
}
