% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{MouseIntestineVisHD}
\alias{MouseIntestineVisHD}
\title{Retrieve 10x-published mouse intestine sample, Visium HD 3.0.0}
\usage{
MouseIntestineVisHD(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve 10x-published mouse intestine sample, Visium HD 3.0.0
}
\note{
From 
`https://www.10xgenomics.com/datasets/visium-hd-cytassist-gene-expression-libraries-of-mouse-intestine`
It takes at least a minute for this function to return, as there is
considerable decompression and reassembly once the basic resource has
been retrieved from cache.
}
