% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\name{MulticancerSteinbock}
\alias{MulticancerSteinbock}
\title{Retrieve spatial mass cytometry experiments for four cancers}
\usage{
MulticancerSteinbock(target = tempfile())
}
\arguments{
\item{target}{character(1) defaults to tempfile().  Set to
a different folder for persistent Zarr store.}
}
\description{
Retrieve spatial mass cytometry experiments for four cancers
}
\note{
From spatialdata archive 
citing `https://www.nature.com/articles/s41596-023-00881-0`.
See also `https://bodenmillergroup.github.io/steinbock/latest/`.
}
\examples{
MulticancerSteinbock()
}
