% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\name{neanderthal_extra}
\alias{neanderthal_extra}
\title{Download Environmental Raster for neanderthal}
\usage{
neanderthal_extra()
}
\value{
SpatRaster object.
}
\description{
Downloads and reads the environmental raster associated with the \link{neanderthal} dataset from the \href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository. Writes the file \code{neanderthal_env.tif} to the working directory and returns it as a \code{spatRaster} object.
}
\seealso{
Other neanderthal: 
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}}
}
\concept{neanderthal}
