% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{path_to_10x_xen_demo}
\alias{path_to_10x_xen_demo}
\title{provide path to a zip file from 10x genomics for Xenium platform}
\usage{
path_to_10x_xen_demo(zipname = "Xenium_V1_human_Breast_2fov_outs.zip")
}
\arguments{
\item{zipname}{character(1)}
}
\description{
provide path to a zip file from 10x genomics for Xenium platform
}
\examples{
p1 = path_to_10x_xen_demo()
td = tempfile()
dir.create(td)
unzip(p1, exdir=td)
target = tempfile()
use_sdio("xenium")(td)$write(target)
br2fov = SpatialData::readSpatialData(target)
}
