% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\name{plantae_east}
\alias{plantae_east}
\title{Subset \code{plantae} to Eastern Hemisphere Ecoregions}
\usage{
plantae_east()
}
\value{
sf data.frame with non-American ecoregions.
}
\description{
Returns a subset of the \link{plantae} dataset filtered to non-American ecoregions
(\code{ecoregion_continent != "Americas"}) with \code{richness_species}
(overall plant species richness) as the only response variable. All 84
predictor variables and identifier columns are retained.
}
\examples{
east <- plantae_east()
nrow(east)
names(east)
}
\seealso{
Other plantae: 
\code{\link{plantae}},
\code{\link{plantae_extra}()},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}},
\code{\link{plantae_west}()}
}
\concept{plantae}
