% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\docType{data}
\name{plantae_predictors}
\alias{plantae_predictors}
\title{Predictor variable names for plantae dataset}
\format{
A character vector of length 84.
}
\usage{
data(plantae_predictors)
}
\description{
Character vector of 84 predictor variable names from \code{plantae}, covering
sampling bias (1), geographic/geometric (10), human impact (3), climate change (2),
landcover (3), fragmentation (19), atmospheric (13), precipitation (8), temperature (11),
soil properties (6), soil temperature (4), and NDVI (4).
}
\seealso{
Other plantae: 
\code{\link{plantae}},
\code{\link{plantae_east}()},
\code{\link{plantae_extra}()},
\code{\link{plantae_responses}},
\code{\link{plantae_west}()}
}
\concept{plantae}
\keyword{datasets}
