% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\docType{data}
\name{plantae_responses}
\alias{plantae_responses}
\title{Response variable names for the dataset \code{plantae}}
\format{
A character vector of length 53.
}
\usage{
data(plantae_responses)
}
\description{
Character vector containing the names of the 53 response variables in \code{plantae}.
}
\examples{
data(plantae_responses)
plantae_responses
}
\seealso{
Other plantae:
\code{\link{plantae}},
\code{\link{plantae_east}},
\code{\link[=plantae_extra]{plantae_extra()}},
\code{\link{plantae_predictors}},
\code{\link{plantae_west}}
}
\concept{plantae}
\keyword{datasets}
