% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\name{plantae_west}
\alias{plantae_west}
\title{Subset \code{plantae} to Western Hemisphere Ecoregions}
\usage{
plantae_west()
}
\value{
sf data.frame.
}
\description{
Returns a subset of the \link{plantae} dataset filtered to American ecoregions
(\code{ecoregion_continent == "Americas"}) with \code{richness_species}
(overall plant species richness) as the only response variable. All 84
predictor variables and identifier columns are retained.
}
\examples{
west <- plantae_west()
nrow(west)
names(west)
}
\seealso{
Other plantae: 
\code{\link{plantae}},
\code{\link{plantae_east}()},
\code{\link{plantae_extra}()},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}}
}
\concept{plantae}
