% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\docType{data}
\name{plants_predictors}
\alias{plants_predictors}
\title{Predictor variable names for plant richness examples}
\format{
A character vector of length 17.
}
\usage{
data(plants_predictors)
}
\description{
Character vector of predictor variable names from \code{plants} (columns 3 to 19).
}
\seealso{
Other plants: 
\code{\link{plants}},
\code{\link{plants_response}}
}
\concept{plants}
\keyword{datasets}
