% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_utils.R
\name{plotCoordGraph}
\alias{plotCoordGraph}
\title{CT graph viz.}
\usage{
plotCoordGraph(g, cex = 0.6)
}
\arguments{
\item{g}{base R graph; extracted with \code{\link{CTgraph}}.}

\item{cex}{scalar numeric; controls fontsize of node labels.}
}
\description{
given a \code{graphNEL} instance, nodes with \code{nchar>max} are
split and hyphenated at character position \code{floor(nchar/fac)}.
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

g <- CTgraph(x)
plotCoordGraph(g, cex=0.6)

}
