% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotImage}
\alias{plotImage}
\alias{plotSpatialData}
\alias{plotImage,SpatialData-method}
\title{\code{SpatialData} image viz.}
\usage{
plotSpatialData()

\S4method{plotImage}{SpatialData}(x, i = 1, j = 1, k = NULL)
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i}{element to use from a given layer.}

\item{j}{name of target coordinate system.}

\item{k}{index of the scale of an image; by default (NULL), will auto-select 
scale in order to minimize memory-usage and blurring for a target size of 
800 x 800px; use Inf to plot the lowest resolution available.}
}
\value{
ggplot
}
\description{
...
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

ms <- lapply(seq(3), \(.) 
  plotSpatialData() +
  plotImage(x, i=2, k=.))
patchwork::wrap_plots(ms)
}
