% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLabel.R
\name{plotLabel}
\alias{plotLabel}
\alias{plotLabel,SpatialData-method}
\title{\code{SpatialData} label viz.}
\usage{
\S4method{plotLabel}{SpatialData}(
  x,
  i = 1,
  c = NULL,
  a = 0.5,
  pal = c("red", "green"),
  nan = NA,
  assay = 1
)
}
\arguments{
\item{x}{\code{SpatialData} object.}

\item{i}{character string or index; the label element to plot.}

\item{c}{the default, NULL, gives a binary image of whether or not 
a given pixel is non-zero; alternatively, a character string specifying
a \code{colData} column or row name in a \code{table} annotating \code{i}.}

\item{a}{scalar numeric in [0, 1]; alpha value passed to \code{geom_tile}.}

\item{pal}{character vector; color for discrete/continuous values
(interpolated automatically when insufficient values are provided).}

\item{nan}{character string; color for missing values (hidden by default).}

\item{assay}{character string; in case of \code{c} denoting a row name,
specifies which \code{assay} data to use (see \code{\link{valTable}}).}
}
\description{
\code{SpatialData} label viz.
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, anndataR=TRUE)

i <- "blobs_labels"
p <- plotSpatialData()

# simple binary image
p + plotLabel(x, i)

# coloring by 'colData'
p + plotLabel(x, i, "instance_id")

# coloring by 'assay' data
p + plotLabel(x, i, "channel_1_sum")
}
