% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoint.R
\name{plotPoint}
\alias{plotPoint}
\alias{plotPoint,SpatialData-method}
\alias{plotPoint,PointFrame-method}
\title{\code{SpatialData} point viz.}
\usage{
\S4method{plotPoint}{SpatialData}(x, i = 1, c = NULL, s = 1, a = 1, assay = 1)

\S4method{plotPoint}{PointFrame}(x, c = NULL, s = 1, a = 1)
}
\arguments{
\item{x}{\code{SpatialData} object or \code{PointFrame}.}

\item{i}{character string or index; the label element to plot.}

\item{c}{character string giving a color; alternatively, may specify a
\code{colData} column or row name in a \code{table} annotating \code{i}.}

\item{s}{scalar numeric; size value passed to \code{geom_point}.}

\item{a}{scalar numeric in [0, 1]; alpha value passed to \code{geom_point}.}

\item{assay}{character string; in case of \code{c} denoting a row name,
specifies which \code{assay} data to use (see \code{\link{valTable}})
(ignored when \code{x} is a \code{PointFrame}).}
}
\description{
\code{SpatialData} point viz.
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, anndataR=TRUE)

i <- "blobs_points"
p <- plotSpatialData()

# mock up a 'table'
f <- list(
  numbers=\(n) runif(n),
  letters=\(n) sample(letters, n, TRUE))
x <- setTable(x, i, f)

p + plotPoint(x, i) # simple dots
p + plotPoint(x, i, "numbers") # discrete coloring
p + plotPoint(x, i, "letters") # continuous coloring

}
