% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotting}
\alias{plotting}
\alias{plotImage}
\alias{plotLabel}
\alias{plotShape}
\alias{plotPoint}
\alias{pile}
\alias{plotImage,SpatialData-method}
\alias{plotLabel,SpatialData-method}
\alias{plotShape,SpatialData-method}
\alias{plotPoint,SpatialData-method}
\alias{pile,list-method}
\alias{pile,ANY-method}
\title{`SpatialData` plotting}
\usage{
\S4method{plotImage}{SpatialData}(x, ..., i = 1, coord = 1, width = 400)

\S4method{plotLabel}{SpatialData}(
  x,
  ...,
  i = 1,
  coord = 1,
  fill = NULL,
  assay = "X",
  alpha = 0.5,
  pal = hcl.colors(9, "Spectral"),
  width = NULL
)

\S4method{plotShape}{SpatialData}(x, ..., i = 1, coord = 1, assay = "X", n = NULL)

\S4method{plotPoint}{SpatialData}(x, ..., i = 1, coord = 1)

\S4method{pile}{list}(...)

\S4method{pile}{ANY}(...)
}
\arguments{
\item{x}{A \code{\link{SpatialData}} object.}

\item{...}{Additional graphical parameters passed to different methods.}

\item{i}{Scalar integer or character string 
specifying which element to plot.}

\item{coord}{Scalar integer or character string
specifying the target coordinate system.}

\item{width}{Scalar integer specifying the output's width (in pixel).}

\item{fill, alpha}{Graphical parameters.}

\item{assay}{Scalar integer or character string
specifying the corresponding `table` data to use.}

\item{pal}{Color palette to use.}

\item{n}{Scalar integer; for \code{plotShapes} when rendering circles,
specifies the number of vertices to compute for each. If \code{NULL},
will automatically scale inversely with the number of instances.}
}
\value{
\code{ggplot2}
}
\description{
...
}
\examples{
dir <- file.path("extdata", "blobs")
dir <- system.file(dir, package="SpatialData")
spd <- readSpatialData(dir)

# element-wise
i <- plotImage(spd)
l <- plotLabel(spd)
s <- plotShape(spd)
p <- plotPoint(spd)

# layered
pile(i, l, s, p) # or...
pile(list(i, l))
}
\author{
Helena L. Crowell
}
