% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quercus.R
\name{quercus_extra}
\alias{quercus_extra}
\title{Download Environmental Raster for quercus}
\usage{
quercus_extra()
}
\value{
SpatRaster object.
}
\description{
Downloads and reads the environmental raster associated with the \link{quercus} dataset from the \href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository. Writes the file \code{quercus_env.tif} to the working directory, and returns it as a \code{spatRaster} object.
}
\seealso{
Other quercus:
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}}
}
\concept{quercus}
