% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quercus.R
\docType{data}
\name{quercus_predictors}
\alias{quercus_predictors}
\title{Predictor variable names for quercus dataset}
\format{
A character vector of length 31.
}
\usage{
data(quercus_predictors)
}
\description{
Character vector of 31 predictor variable names from \link{quercus}, covering
bioclimatic (17), NDVI (4), solar radiation (4), land cover (3), topography (2),
and human footprint (1).
}
\seealso{
Other quercus: 
\code{\link{quercus}},
\code{\link{quercus_extra}()},
\code{\link{quercus_response}}
}
\concept{quercus}
\keyword{datasets}
