% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quercus.R
\docType{data}
\name{quercus_response}
\alias{quercus_response}
\title{Response variable name for the dataset \code{quercus}}
\format{
A character string of length 1.
}
\usage{
data(quercus_response)
}
\description{
Character string with the name of the response variable in \link{quercus}.
}
\examples{
data(quercus_response)
quercus_response
}
\seealso{
Other quercus:
\code{\link{quercus}},
\code{\link[=quercus_extra]{quercus_extra()}},
\code{\link{quercus_predictors}}
}
\concept{quercus}
\keyword{datasets}
