% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query,SpatialData-method}
\alias{query,ImageArray-method}
\alias{query,LabelArray-method}
\alias{query,ShapeFrame-method}
\alias{query,PointFrame-method}
\title{spatial queries}
\usage{
\S4method{query}{SpatialData}(x, j = NULL, ...)

\S4method{query}{ImageArray}(x, ...)

\S4method{query}{LabelArray}(x, ...)

\S4method{query}{ShapeFrame}(x, ...)

\S4method{query}{PointFrame}(x, j, ...)
}
\arguments{
\item{x}{\code{SpatialData} element.}

\item{j}{scalar character or integer; index or name of coordinate space.}

\item{...}{optional arguments passed to and from other methods.}
}
\value{
same as input
}
\description{
spatial queries
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")
sd <- readSpatialData(zs, tables=FALSE)

image(sd, "box") <- query(image(sd), xmin=0, xmax=30, ymin=30, ymax=50)

image(sd)
image(sd, "box")
}
