% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArray.R
\name{readArray}
\alias{readArray}
\title{Read `images/labels` element}
\usage{
readArray(path = ".", resolution = "0", ...)
}
\arguments{
\item{path}{A character string specifying
a .zarray or .zattrs file-containing directory.}

\item{resolution}{A charactering specifiying
the image resolution (pyramid level) to read in.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
\code{\link{ImageArray}}
}
\description{
...
}
\examples{
path <- system.file("extdata", "blobs", package="SpatialData")
(ia <- readArray(file.path(path, "images", "blobs_image")))
(la <- readArray(file.path(path, "labels", "blobs_labels")))

}
\author{
Helena L. Crowell
}
