% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImage.R
\name{readImage}
\alias{readImage}
\title{Read as `ImageArray`}
\usage{
readImage(path = ".", resolution = "0", ...)
}
\arguments{
\item{path}{A character string specifying
a .zarray or .zattrs file-containing directory.}

\item{resolution}{A charactering specifiying
the image resolution (pyramid level) to read in.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\description{
...
}
\examples{
path <- "extdata/mibitof/images/point8_image/0"
path <- system.file(path, package = "SpatialData")
(ia <- readImage(path))

}
