% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readPoints}
\alias{readPoints}
\title{Read `points` element}
\usage{
readPoints(path, ...)
}
\arguments{
\item{path}{A character string specifying
a .parquet file-containing directory.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
Arrow \code{\link{Dataset}}
}
\description{
...
}
\examples{
dir <- "extdata/blobs/points/blobs_points"
dir <- system.file(dir, package = "SpatialData")
(pf <- readPoints(dir))

}
\author{
Tim Treis
}
