% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readShapes.R
\name{readShapes}
\alias{readShapes}
\title{Read `shapes` element}
\usage{
readShapes(path, ...)
}
\arguments{
\item{path}{A character string specifying
the path to a \code{shapes/} subdirectory.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
\code{\link[S4Vectors]{DataFrame}}
}
\description{
...
}
\examples{
path <- file.path("extdata", "raccoon", "shapes", "circles")
path <- system.file(path, package="SpatialData")
(sf <- readShapes(path))
as.data.frame(sf)

}
\author{
Tim Treis, Helena L. Crowell
}
