% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readTable}
\alias{readTable}
\title{`SingleCellExperiment` from `AnnData`-Zarr}
\usage{
readTable(path)
}
\arguments{
\item{path}{A character string specifying
the path to a `table/` subdirectory.}
}
\value{
\code{SingleCellExperiment}
}
\description{
...
}
\examples{
dir <- file.path("extdata", "blobs", "table", "table")
dir <- system.file(dir, package="SpatialData")
(sce <- readTable(dir))

}
\author{
Constantin Ahlmann-Eltze
}
