% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdFrame.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pull}
\alias{select}
\alias{mutate}
\alias{filter}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:pull]{pull()}}, \code{\link[dplyr:select]{select()}}}
}}

