% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{SD-miscellaneous}
\alias{SD-miscellaneous}
\alias{show,SpatialData-method}
\alias{show,ZarrArray-method}
\alias{show,ImageArray-method}
\alias{show,LabelArray-method}
\alias{show,ShapeFrame-method}
\alias{show,PointFrame-method}
\title{Miscellaneous `SpatialData` methods}
\usage{
\S4method{show}{SpatialData}(object)

\S4method{show}{ZarrArray}(object)

\S4method{show}{ImageArray}(object)

\S4method{show}{LabelArray}(object)

\S4method{show}{ShapeFrame}(object)

\S4method{show}{PointFrame}(object)
}
\arguments{
\item{object}{\code{\link{SpatialData}} or \code{\link{ImageArray}} object.}
}
\value{
\code{NULL}
}
\description{
Miscellaneous methods for the \code{\link{SpatialData}}
and \code{\link{ImageArray}} classes that do not fit
into any other documentation category such as,
for example, show methods.
}
\examples{
dir <- system.file("extdata", "raccoon", package="SpatialData")
(img <- readArray(file.path(dir, "images", "raccoon")))
(spd <- readSpatialData(dir))
}
\author{
Helena L. Crowell
}
