% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sd_plot}
\alias{sd_plot}
\alias{sd_plot_image}
\title{Plotting `SpatialData`}
\usage{
sd_plot_image(
  x,
  i = 1,
  k = NULL,
  c = NULL,
  ch = NULL,
  cl = NULL,
  w = 800,
  h = 800
)

sd_plot()
}
\description{
Plotting `SpatialData`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

sd_plot() + sd_plot_image(sd)

pal <- c("cyan", "magenta", "gold")
sd_plot() + sd_plot_image(sd, c=pal)
}
