% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-image.R
\name{sd_plot_image}
\alias{sd_plot_image}
\title{Plot `ImageArray`}
\usage{
sd_plot_image(
  x,
  i = 1,
  k = NULL,
  c = NULL,
  ch = NULL,
  cl = NULL,
  w = 800,
  h = 800
)
}
\description{
Plot `ImageArray`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

sd_plot() + sd_plot_image(sd)

# custom colors
pal <- c("cyan", "magenta", "gold")
sd_plot() + sd_plot_image(sd, c=pal)

}
