% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-point.R
\name{sd_plot_point}
\alias{sd_plot_point}
\title{Plot `PointFrame`}
\usage{
sd_plot_point(x, i, c, a, ...)
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i}{scalar integer or string; 
specifies which \code{points} to plot.}

\item{c}{string; can be a color name or 
\code{PointFrame} column to color by.}

\item{a}{assay to use when coloring by \code{tables}.}

\item{...}{(optional) aesthetics passed to \code{geom_point}.}
}
\value{
\code{ggplot}
}
\description{
Plot `PointFrame`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

sd_plot() + sd_plot_point(sd, c="x")
sd_plot() + sd_plot_point(sd, c="pink", size=1)

}
