% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-shape.R
\name{sd_plot_shape}
\alias{sd_plot_shape}
\title{Plot `ShapeFrame`}
\usage{
sd_plot_shape(x, i = 1, c = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i}{scalar integer or string; 
specifies which \code{shapes} to plot.}

\item{c}{string; can be a color name,
or TRUE to color by shape identifier.}

\item{...}{(optional) aesthetics passed to \code{geom_sf}.}
}
\value{
\code{ggplot}
}
\description{
Plot `ShapeFrame`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

sd_plot() + sd_plot_shape(sd, 1) # point
sd_plot() + sd_plot_shape(sd, 3) # polygon
sd_plot() + sd_plot_shape(sd, 2) # multi-polygon

# aesthetics
sd_plot() + sd_plot_shape(sd, 3, 
  fill=NA, c="red", linewidth=1)

}
