% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdArray.R
\name{sdArray}
\alias{sdArray}
\alias{data,sdArray-method}
\alias{dim,sdArray-method}
\alias{length,sdArray-method}
\alias{data_type,sdArray-method}
\alias{data_type,DelayedArray-method}
\title{Methods for `ImageArray` and `LabelArray` class}
\usage{
\S4method{data}{sdArray}(x, k = 1)

\S4method{dim}{sdArray}(x)

\S4method{length}{sdArray}(x)

\S4method{data_type}{sdArray}(x)

\S4method{data_type}{DelayedArray}(x)
}
\arguments{
\item{x}{\code{ImageArray} or  \code{LabelArray}}

\item{k}{scalar index specifying which scale to extract.}
}
\value{
\code{ImageArray}
}
\description{
Methods for `ImageArray` and `LabelArray` class
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")

pa <- list.dirs(
  file.path(zs, "images"), 
  recursive=FALSE, full.names=TRUE)

(x <- readImage(pa[2]))  

channels(x)
data_type(x)
dim(data(x, 1))   # highest res.
dim(data(x, Inf)) # lowest res.

rgb <- apply(
  data(x, 1), c(2, 3), 
  \(.) rgb(.[1], .[2], .[3]))
plot(
  row(rgb), col(rgb), col=rgb, 
  pch=15, asp=1, ylim=c(ncol(rgb), 0))

}
