% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdFrame.R
\name{sdFrame}
\alias{sdFrame}
\alias{dim,sdFrame-method}
\alias{names,sdFrame-method}
\alias{[[,sdFrame,ANY,ANY-method}
\alias{.DollarNames.ShapeFrame}
\alias{$,ShapeFrame-method}
\alias{as.data.frame,sdFrame-method}
\alias{pull.sdFrame}
\alias{select.sdFrame}
\alias{mutate.sdFrame}
\alias{filter.sdFrame}
\title{The `sdFrame` class}
\usage{
\S4method{length}{sdFrame}(x)

\S4method{dim}{sdFrame}(x)

\S4method{names}{sdFrame}(x)

\S4method{[[}{sdFrame,ANY,ANY}(x, i, j, ...)

\method{.DollarNames}{ShapeFrame}(x, pattern = "")

\S4method{$}{ShapeFrame}(x, name)

\S4method{as.data.frame}{sdFrame}(x)

\method{pull}{sdFrame}(.data, ...)

\method{select}{sdFrame}(.data, ...)

\method{mutate}{sdFrame}(.data, ...)

\method{filter}{sdFrame}(.data, ...)
}
\arguments{
\item{x}{\code{PointFrame}}

\item{i, j}{indices for subsetting (see \code{?base::Extract}).}

\item{...}{optional arguments passed to and from other methods.}

\item{name}{character string for extraction (see \code{?base::`$`}).}

\item{data}{\code{arrow}-derived table for on-disk,
\code{data.frame} for in-memory representation.}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary
content describing the overall object.}

\item{drop}{ignored.}
}
\value{
\code{PointFrame}
}
\description{
The \code{PointFrame} class stores \code{SpatialData} elements from its
\code{"points"} layers. These are represented as \code{\link[arrow]{Table}}
(\code{data} slot) associated with .zattrs stored as \code{\link{Zattrs}}
(\code{meta} slot); a list of \code{metadata} stores other arbitrary info.

Currently defined methods (here, \code{x} is a \code{PointFrame}):
\itemize{
\item \code{data/meta(x)} to access underlying \code{Table/Zattrs}
\item \code{names(x)} returns the underlying table's column names
\item \code{dim(x)} returns the dimensions of \code{data(x)}
\item \code{`$`,`[[`} directly access columns of \code{data(x)}
\item \code{filter,select} to subset rows/columns à la \code{dplyr}
\item \code{as.data.frame} to coerce \code{x} to a \code{data.frame}
}
}
\examples{
library(SpatialData.data)
zs <- get_demo_SDdata("merfish")
x <- file.path(zs, "points", "single_molecule")
(p <- readPoint(x))

head(as.data.frame(data(p)))
(q <- dplyr::filter(p, cell_type == "VISp_wm"))

}
