% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{sdImage}
\alias{sdImage}
\alias{channels}
\alias{channels,sdImage-method}
\alias{channels,ANY-method}
\alias{[,sdImage,ANY,ANY,ANY-method}
\title{The `sdImage` class}
\usage{
sdImage(data = list(), meta = Zattrs(), metadata = list(), ...)

\S4method{channels}{sdImage}(x, ...)

\S4method{channels}{ANY}(x, ...)

\S4method{[}{sdImage,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)
}
\arguments{
\item{data}{list of \code{\link[ZarrArray]{ZarrArray}}s}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{sdImage}}

\item{i, j}{indices specifying elements to extract.}

\item{k}{scalar index specifying which scale to extract.}

\item{drop}{ignored.}
}
\value{
\code{sdImage}
}
\description{
The `sdImage` class
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")

pa <- list.dirs(
  file.path(zs, "images"), 
  recursive=FALSE, full.names=TRUE)

# simple
readImage(pa[1])

# multi-scale
(x <- readImage(pa[2]))

dim(data(x, 1))   # highest res.
dim(data(x, Inf)) # lowest res.

rgb <- apply(
  data(x, 1), c(2, 3), 
  \(.) rgb(.[1], .[2], .[3]))
plot(
  row(rgb), col(rgb), col=rgb, 
  pch=15, asp=1, ylim=c(ncol(rgb), 0))

}
