% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{getTable}
\alias{getTable}
\title{\code{SpatialData} tables}
\usage{
getTable(x, ...)
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i}{scalar character or integer; name of the 
element for which a `table` should be retrieved.}

\item{drop}{logical; should observations (columns) 
that don't belong to \code{i} be filtered out?}
}
\description{
\code{SpatialData} tables
}
