% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFrame.R, R/sdFrame.R
\name{ShapeFrame}
\alias{ShapeFrame}
\alias{geom_type}
\alias{geom_type,ShapeFrame-method}
\alias{[,sdFrame,ANY,ANY,ANY-method}
\title{The `ShapeFrame` class}
\usage{
ShapeFrame(
  data = data.frame(),
  meta = Zattrs(type = "frame"),
  metadata = list(),
  ...
)

\S4method{geom_type}{ShapeFrame}(x)

\S4method{[}{sdFrame,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{data}{\code{duckspatial_df} for on-disk representation,
a 3-column \code{data.frame} (with columns \code{x}, \code{y} and
\code{id}) with vertices of polygons, or any object that can be passed
to \code{\link[duckspatial]{as_duckspatial_df}}.}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary
content describing the overall object.}

\item{...}{optional arguments passed to and from other methods.}

\item{x}{\code{ShapeFrame}}

\item{i, j}{indices specifying elements to extract.}

\item{drop, pattern}{ignored.}

\item{name}{character string for extraction (see \code{?base::`$`}).}
}
\value{
\code{ShapeFrame}
}
\description{
The `ShapeFrame` class
}
\examples{
library(SpatialData.data)
zs <- get_demo_SDdata("merfish")

y <- file.path(zs, "shapes", "cells")
(s <- readShape(y))
plot(sf::st_as_sf(data(s)), cex=0.2)

y <- file.path(zs, "shapes", "anatomical")
(s <- readShape(y))
plot(sf::st_as_sf(data(s)), cex=0.2)

}
