% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{spatialdata-io}
\alias{spatialdata-io}
\alias{sd_make}
\alias{sdio_list}
\title{Writing .zarr}
\usage{
sd_make(platform, srcdir, dest, env)

sdio_list(env)
}
\arguments{
\item{platform}{string; must be an element of `sdio_list()` output}

\item{srcdir}{string; path to folder holding manufacturer output files}

\item{dest}{string; path to a desired destination for .zarr representation}

\item{env}{(optional) string passed to `reticulate::use_condaenv()`;
can be used to specify a conda environment with the 
`spatialdata_io` Python module already installed.
.}
}
\value{
none, but a .zarr store is created 
  in the specified \code{dest} location.
}
\description{
Use Python's 'spatialdata-io' to transform manufacturer 
output to .zarr with specific folder structure.
}
\examples{
# unzip flat files
pa <- SpatialData.data:::.path_to_10x_xen_demo()
dir.create(td <- tempfile())
unzip(pa, exdir=td)

# read & write to .zarr w/ 'spatialdata-io'
out <- tempfile()
sd_make("xenium", 
  srcdir=td, dest=out,
  env="spatialdata-io")
(sd <- readSpatialData(out))

}
