% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SpatialData.R, R/methods.R
\docType{class}
\name{SpatialData-class}
\alias{SpatialData-class}
\alias{.SpatialData}
\alias{SpatialData}
\alias{data}
\alias{meta}
\alias{layer}
\alias{element}
\alias{element<-}
\alias{image}
\alias{label}
\alias{point}
\alias{shape}
\alias{table}
\alias{images}
\alias{labels}
\alias{points}
\alias{shapes}
\alias{tables}
\alias{image<-}
\alias{label<-}
\alias{point<-}
\alias{shape<-}
\alias{table<-}
\alias{images<-}
\alias{labels<-}
\alias{points<-}
\alias{shapes<-}
\alias{tables<-}
\alias{imageNames}
\alias{labelNames}
\alias{pointNames}
\alias{shapeNames}
\alias{tableNames}
\alias{imageNames<-}
\alias{labelNames<-}
\alias{pointNames<-}
\alias{shapeNames<-}
\alias{tableNames<-}
\alias{[[<-,SpatialData,character,ANY-method}
\alias{[[<-,SpatialData,numeric,ANY-method}
\alias{$,SpatialData-method}
\alias{$<-,SpatialData-method}
\alias{[[,SpatialData,numeric,ANY-method}
\alias{[[,SpatialData,character,ANY-method}
\alias{data,SpatialDataElement-method}
\alias{meta,SpatialDataElement-method}
\alias{[,SpatialData,ANY,ANY,ANY-method}
\alias{rownames,SpatialData-method}
\alias{colnames,SpatialData-method}
\alias{layer,SpatialData,character-method}
\alias{layer,SpatialData,ANY-method}
\alias{element,SpatialData,character-method}
\alias{element,SpatialData,numeric-method}
\alias{element,SpatialData,missing-method}
\alias{element,SpatialData,ANY-method}
\alias{element<-,SpatialData,character-method}
\alias{images,SpatialData-method}
\alias{labels,SpatialData-method}
\alias{points,SpatialData-method}
\alias{shapes,SpatialData-method}
\alias{tables,SpatialData-method}
\alias{[[<-,SpatialData,ANY,ANY-method}
\title{The `SpatialData` class}
\usage{
SpatialData(
  images = list(),
  labels = list(),
  points = list(),
  shapes = list(),
  tables = list()
)

\S4method{$}{SpatialData}(x, name)

\S4method{$}{SpatialData}(x, name) <- value

\S4method{[[}{SpatialData,numeric,ANY}(x, i, j, ...)

\S4method{[[}{SpatialData,character,ANY}(x, i, j, ...)

\S4method{data}{SpatialDataElement}(x, k = 1, ...)

\S4method{meta}{SpatialDataElement}(x)

\S4method{[}{SpatialData,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{rownames}{SpatialData}(x)

\S4method{colnames}{SpatialData}(x)

\S4method{layer}{SpatialData,character}(x, i)

\S4method{layer}{SpatialData,ANY}(x, i)

\S4method{element}{SpatialData,character}(x, i)

\S4method{element}{SpatialData,numeric}(x, i)

\S4method{element}{SpatialData,missing}(x, i)

\S4method{element}{SpatialData,ANY}(x, i)

\S4method{element}{SpatialData,character}(x, i) <- value

\S4method{images}{SpatialData}(x)

\S4method{labels}{SpatialData}(object)

\S4method{points}{SpatialData}(x)

\S4method{shapes}{SpatialData}(x)

\S4method{tables}{SpatialData}(x)

\S4method{[[}{SpatialData,character,ANY}(x, i) <- value

\S4method{[[}{SpatialData,numeric,ANY}(x, i) <- value

\S4method{[[}{SpatialData,ANY,ANY}(x, i) <- value
}
\arguments{
\item{images}{list of \code{\link{SpatialDataImage}}s}

\item{labels}{list of \code{\link{SpatialDataLabel}}s}

\item{points}{list of \code{\link{SpatialDataPoint}}s}

\item{shapes}{list of \code{\link{SpatialDataShape}}s}

\item{tables}{list of \code{SingleCellExperiment}s}

\item{x, object}{\code{SpatialData} object.}

\item{name}{character string for extraction (see \code{?base::`$`}).}

\item{value}{(list of) element(s) with layer-compliant object(s), 
or NULL/\code{list()} to remove an element/layer completely; 
for \code{element<-}, a single \code{SpatialDataElement}
of the same class as \code{element(x, i)}.}

\item{i, j}{character string, scalar or vector of indices
specifying the element to extract from a given layer.}

\item{...}{optional arguments passed to and from other methods.}

\item{k}{scalar index specifying which scale to use; 
\code{Inf} to use lowest available resolution.}

\item{drop}{ignored.}
}
\value{
\code{SpatialData}
}
\description{
\code{SpatialData} provides an R interface to Python's \code{spatialdata},
which enables the representation of diverse spatial omics datasets using 
the OME-NGFF (Next Generation File Format) standard. In R, 
\itemize{
\item images and labels are \code{ZarrArray}s (\code{Rarr} package).
\item points and shapes are managed using \code{duckspatial} tables.
\item tables are \code{SingleCellExperiment}s (read with \code{anndataR}).}
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="spatialdataR")
(x <- readSpatialData(x))

# subsetting
# layers are taken in order of appearance
# (images, labels, points, shapes, tables)
x[-4] # drop layer
x[4, -2] # drop element
x["shapes", c(1, 3)] # subset layer
x[c(1, 2), list(1, c(1, 2))] # multiple

}
