% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdFrame.R
\name{SpatialDataFrame}
\alias{SpatialDataFrame}
\alias{SpatialDataPoint}
\alias{SpatialDataShape}
\alias{geom_type}
\alias{length,SpatialDataFrame-method}
\alias{dim,SpatialDataFrame-method}
\alias{names,SpatialDataFrame-method}
\alias{as.data.frame,SpatialDataFrame-method}
\alias{geom_type,SpatialDataShape-method}
\alias{pull.SpatialDataFrame}
\alias{select.SpatialDataFrame}
\alias{mutate.SpatialDataFrame}
\alias{filter.SpatialDataFrame}
\alias{[[,SpatialDataFrame,ANY,ANY-method}
\alias{$,SpatialDataPoint-method}
\alias{.DollarNames.SpatialDataShape}
\alias{$,SpatialDataShape-method}
\alias{[,SpatialDataFrame,ANY,ANY,ANY-method}
\title{\code{SpatialDataFrame}}
\usage{
SpatialDataPoint(
  data = NULL,
  meta = SpatialDataAttrs(type = "frame"),
  version = point(sdFormat(0.1)),
  metadata = list(),
  ik = NULL,
  fk = NULL,
  ...
)

SpatialDataShape(
  data = NULL,
  meta = SpatialDataAttrs(type = "frame"),
  version = shape(sdFormat(0.1)),
  metadata = list(),
  ...
)

\S4method{length}{SpatialDataFrame}(x)

\S4method{dim}{SpatialDataFrame}(x)

\S4method{names}{SpatialDataFrame}(x)

\S4method{as.data.frame}{SpatialDataFrame}(x)

\S4method{geom_type}{SpatialDataShape}(x)

\method{pull}{SpatialDataFrame}(.data, ...)

\method{select}{SpatialDataFrame}(.data, ...)

\method{mutate}{SpatialDataFrame}(.data, ...)

\method{filter}{SpatialDataFrame}(.data, ...)

\S4method{[[}{SpatialDataFrame,ANY,ANY}(x, i, j, ...)

\S4method{$}{SpatialDataPoint}(x, name)

\method{.DollarNames}{SpatialDataShape}(x, pattern = "")

\S4method{$}{SpatialDataShape}(x, name)

\S4method{[}{SpatialDataFrame,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{data}{\code{duckspatial_df} for on-disk representation,
or a \code{data.frame} to be converted.}

\item{meta}{\code{\link{SpatialDataAttrs}}}

\item{metadata}{optional list of arbitrary
content describing the overall object.}

\item{ik, fk}{character string specifying "instance_/feature_key" 
of the spatialdata_attrs; used to match observations/features.}

\item{...}{optional arguments passed to and from other methods.}

\item{x, .data}{\code{SpatialDataFrame}}

\item{i, j}{indices for subsetting (see \code{?base::Extract}).}

\item{name}{character string for extraction (see \code{?base::`$`}).}

\item{drop, pattern}{ignored.}
}
\value{
\code{SpatialDataFrame}
}
\description{
The \code{SpatialDataPoint} and \code{-Shape} classes represent 
elements from a \code{SpatialData}'s \code{points/} and \code{shapes/} 
layers, respectively. In both cases, these  are represented as a 
\code{duckspatial_df} (\code{data} slot), and associated with .zattrs 
represented as \code{\link{SpatialDataAttrs}} (\code{meta} slot); 
a list of \code{metadata} stores other arbitrary info.

Currently defined methods (here, \code{x} is an \code{SpatialDataFrame}):
\itemize{
\item \code{data/meta(x)} access underlying data/.zattrs
\item \code{geom_type(x)} get the shape's type (e.g., POLYGON)
\item \code{names(x)} returns the underlying table's column names
\item \code{dim(x)} returns the dimensions of \code{data(x)}
\item \code{`$`,`[[`} directly access columns of \code{data(x)}
\item \code{filter,select} to subset rows/columns à la \code{dplyr}
\item \code{as.data.frame} to coerce \code{x} to a \code{data.frame}
}
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="spatialdataR")

# points
pa <- list.dirs(
  file.path(zs, "points"), 
  recursive=FALSE, full.names=TRUE)
(x <- readPoint(pa))

y <- filter(x,
  genes == "gene_b",
  instance_id == 7) 
head(as.data.frame(y))

# shapes
pa <- list.dirs(
  file.path(zs, "shapes"), 
  recursive=FALSE, full.names=TRUE)

# circles
(x <- readShape(pa[1]))
length(x)
x$radius

# polygons
(y <- readShape(pa[2]))
df <- as.data.frame(y)
plot(df, col=seq(nrow(df)))

# multi-polygons
(z <- readShape(pa[3]))
df <- as.data.frame(z)
plot(df, col=seq(nrow(df)))
}
