% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{spd_demo_cached_path}
\alias{spd_demo_cached_path}
\title{check cache for demonstration .zarr.zip and return path; 
retrieve data, cache it, return path if not found}
\usage{
spd_demo_cached_path(
  cache = BiocFileCache::BiocFileCache(),
  zipname = "mibitof.zip",
  source
)
}
\arguments{
\item{cache}{defaults to \code{BiocFileCache::BiocFileCache()}}

\item{zipname}{character(1) should be found in Bioconductor OSN bucket}

\item{source}{character(1) one of "biocOSN", "sandbox", "local"}
}
\value{
Returns character(1) path to cached zip file.
}
\description{
check cache for demonstration .zarr.zip and return path; 
retrieve data, cache it, return path if not found
}
