% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\alias{scale}
\alias{rotate}
\alias{translation}
\alias{scale,PointFrame,numeric-method}
\alias{rotate,PointFrame,numeric-method}
\alias{translation,PointFrame,numeric-method}
\alias{scale,ShapeFrame,numeric-method}
\alias{rotate,ShapeFrame,numeric-method}
\alias{translation,ShapeFrame,numeric-method}
\title{Transformations}
\usage{
\S4method{scale}{PointFrame,numeric}(x, t, ...)

\S4method{rotate}{PointFrame,numeric}(x, t, ...)

\S4method{translation}{PointFrame,numeric}(x, t, ...)

\S4method{scale}{ShapeFrame,numeric}(x, t, ...)

\S4method{rotate}{ShapeFrame,numeric}(x, t, ...)

\S4method{translation}{ShapeFrame,numeric}(x, t, ...)
}
\arguments{
\item{x}{\code{SpatialData} element}

\item{t}{transformation data.}

\item{...}{option arguments passed to and from other methods.}

\item{j}{scalar character or numeric; 
name or index of coordinate space.}
}
\description{
Transformations
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# image
y <- x
image(y) <- scale(image(y), c(1, 1, 1/3))
CTpath(image(x), "global")
CTpath(image(y), "global")
  
# point
y <- x
point(y, "rot") <- rotate(point(y), 20)
point(y, "wide") <- scale(point(y), c(1.2, 1))

xy0 <- as.data.frame(point(y))
xy1 <- as.data.frame(point(y, "rot"))
xy2 <- as.data.frame(point(y, "wide"))

plot(xy0[, c(1, 2)], asp=1)
points(xy1[, c(1, 2)], col=2)
points(xy2[, c(1, 2)], col=4)
  
# shape
y <- x
shape(y, "rot") <- rotate(shape(y), 5)
shape(y, "wide") <- scale(shape(y), c(1.2, 1))
shape(y, "left") <- translation(shape(y), c(-5, 0))
y["shapes", c("rot", "wide", "left")]
}
