% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{unzip_merfish_demo}
\alias{unzip_merfish_demo}
\title{unzip cached merfish demo data to specified folder}
\usage{
unzip_merfish_demo(destination, cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{destination}{character(1) a path to a folder that must exist}

\item{cache}{defaults to BiocFileCache::BiocFileCache(), will serve as destination or source}
}
\value{
Returns path to 'merfish.zarr'.
}
\description{
unzip cached merfish demo data to specified folder
}
\examples{
tf <- tempfile()
dir.create(tf)
pa <- unzip_merfish_demo(tf)
dir(pa, full.names=TRUE)

}
