% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{centroids}
\alias{centroids,ANY-method}
\alias{centroids,LabelArray-method}
\alias{centroids,ShapeFrame-method}
\alias{centroids,PointFrame-method}
\title{Utilities}
\usage{
\S4method{centroids}{ANY}(x, ...)

\S4method{centroids}{LabelArray}(x, as = c("data.frame", "matrix"))

\S4method{centroids}{ShapeFrame}(x, as = c("data.frame", "matrix", "list"))

\S4method{centroids}{PointFrame}(x, as = c("data.frame", "list"))
}
\arguments{
\item{x}{a \code{SpatialData} element (any but image)}

\item{as}{determines how results will be returned}
}
\description{
Utilities
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

centroids(label(x))
centroids(shape(x))
centroids(shape(x, 3), "list")

head(centroids(point(x)))
xy <- centroids(point(x), "list")
plot(xy$gene_a, col=a <- "red")
points(xy$gene_b, col=b <- "blue")
legend("topright", legend=names(xy), col=c(a, b), pch=21)
}
