% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{vi_extra}
\alias{vi_extra}
\title{Download extended \code{vi} dataset}
\usage{
vi_extra()
}
\value{
sf data.frame with 30,000 rows and 64 columns (POINT geometry, WGS84).
}
\description{
Downloads and reads the extended version of the \link{vi} dataset (30,000 rows) from the \href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository. Writes the file \code{vi.gpkg} to the working directory, and returns it as an \code{sf} dataframe. See \link{vi} for details on the response variables, predictors, and data sources.
}
\seealso{
Other vi: 
\code{\link{vi}},
\code{\link{vi_predictors}},
\code{\link{vi_responses}},
\code{\link{vi_smol}}
}
\concept{vi}
