% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_predictors}
\alias{vi_predictors}
\title{Predictor variable names for vi dataset}
\format{
A character vector of length 58.
}
\usage{
data(vi_predictors)
}
\description{
Character vector of 58 predictor variable names from \link{vi}, covering
climate classification (3), soil type (1), topography (3), soil water index (4),
soil temperature (4), soil properties (6), solar radiation (4), growing season (4),
temperature (5), rainfall (4), evapotranspiration (4), cloud cover (4), aridity (1),
humidity (4), biogeography (3), and administrative (4).
}
\seealso{
Other vi: 
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_responses}}
}
\concept{vi}
\keyword{datasets}
