% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_predictors_numeric}
\alias{vi_predictors_numeric}
\title{Vector of numeric predictor names in \code{vi}}
\format{
Character vector of length 47.
}
\usage{
data(vi_predictors_numeric)
}
\description{
Vector of numeric predictor names in \code{vi}
}
\seealso{
Other example_data: 
\code{\link{vi}},
\code{\link{vi_predictors}},
\code{\link{vi_predictors_categorical}},
\code{\link{vi_responses}}
}
\concept{example_data}
\keyword{datasets}
