% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_responses}
\alias{vi_responses}
\title{Response variable names for vi dataset}
\format{
A character vector of length 5.
}
\usage{
data(vi_responses)
}
\description{
Character vector containing the names of the 5 response variables in \link{vi},
representing different encodings of NDVI: continuous (\code{vi_numeric}), integer counts
(\code{vi_counts}), binary (\code{vi_binomial}), categorical (\code{vi_categorical}), and factor
(\code{vi_factor}).
}
\seealso{
Other vi: 
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}}
}
\concept{vi}
\keyword{datasets}
