% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_smol}
\alias{vi_smol}
\title{Small version of \code{vi}}
\format{
A data frame with 580 rows and 65 columns.
}
\usage{
data(vi_smol)
}
\description{
Same as dataset \link{vi}, but with only 580 rows.
}
\examples{
data(vi_smol)
colnames(vi_smol)
nrow(vi_smol)
ncol(vi_smol)
}
\seealso{
\link{vi}

Other vi: 
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}},
\code{\link{vi_responses}}
}
\concept{vi}
\keyword{datasets}
