% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someviz.R
\name{viewShape}
\alias{viewShape}
\title{Use geom_sf to view a shapes component}
\usage{
viewShape(sdobj, elem)
}
\arguments{
\item{sdobj}{SpatialData instance}

\item{elem}{character(1) name of a shapes component of sdobj}
}
\description{
Use geom_sf to view a shapes component
}
