% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeImageArray.R
\name{writeImageArray}
\alias{writeImageArray}
\title{Write `ImageArray` to Zarr-array}
\usage{
writeImageArray(image, path, ...)
}
\arguments{
\item{image}{A \code{\link{ImageArray}} specifying the image to be saved.}

\item{path}{A character string specifying the output location.}

\item{...}{Further arguments to be passed to
\code{\link{Rarr}[write_zarr_array]}.}
}
\value{
\code{NULL}
}
\description{
...
}
\examples{
# TODO

}
