% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeSpatialData}
\alias{writeSpatialData}
\alias{writeImage}
\alias{writeLabel}
\alias{writePoint}
\alias{writeShape}
\alias{writeTable}
\title{Writing `SpatialData`}
\usage{
writePoint(x, ...)
}
\arguments{
\item{x}{For \code{writeImage/Label/Point/Shape/Table}, 
path to a \code{SpatialData} element.
For \code{writeSpatialData},
path to a \code{SpatialData}-.zarr store.}

\item{...}{option arguments passed to and from other methods.}
}
\value{
\itemize{
\item{For \code{writeSpatialData}, a \code{SpatialData}.},
\item{For element writers, a \code{ImageArray}, \code{LabelArray}, 
\code{PointFrame}, \code{ShapeFrame}, or \code{SingleCellExperiment}.}}
}
\description{
Writing `SpatialData`
}
