% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zattrs.R
\name{Zattrs}
\alias{Zattrs}
\alias{$,Zattrs-method}
\title{The `Zattrs` class}
\usage{
Zattrs(
  x,
  type = c("array", "frame"),
  label = FALSE,
  trans = NULL,
  ver = "0.3",
  n = 3,
  scale_factors = NULL,
  ...
)

\S4method{$}{Zattrs}(x, name)
}
\arguments{
\item{x}{list extracted from a OME-NGFF compliant .zattrs file.}

\item{type}{character string; either "array" (image/label) or "frame" (point/shape).}

\item{...}{additional attributes (e.g., version, feature_key).}

\item{name}{character string for extraction (see ?base::`$`).}

\item{axes}{list of axes; if NULL, defaults to cyx (array) or xy (frame).}

\item{transformations}{list of transformations; if NULL, defaults to global identity.}
}
\value{
\code{Zattrs}
}
\description{
The `Zattrs` class
}
\details{
When missing \code{x}, \code{Zattrs} will generate a valid object with
default axes (array: cyx, frame: xy) and transformations (identify) 
according to the specified type.
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

(z <- meta(label(x)))

CTname(z)
CTtype(z)
CTdata(z, "scale")

feature_key(point(x))

# constructor
Zattrs(type="frame")
Zattrs(type="array")
Zattrs(type="array", n=7)
Zattrs(type="array", label=TRUE)

}
