% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateImage.R
\name{aggregateImage}
\alias{aggregateImage}
\title{Aggregate to `SingleCellExperiment`}
\usage{
aggregateImage(x, image = 1, label = 1, fun = mean)
}
\arguments{
\item{x}{A \code{\link{SpatialData}} object.}

\item{image, label}{Index or character string specifying
the image/label to use; if a string is provided,
should be one of \code{image/labelNames(x)}.}

\item{fun}{Function to use for aggregation.}
}
\value{
An object of class \code{\link{SingleCellExperiment}}
where rows=image channels and columns=unique labels.
}
\description{
...
}
\examples{
library(ggplot2)
library(SingleCellExperiment)
path <- file.path("extdata", "blobs")
path <- system.file(path, package = "SpatialData")
spd <- readSpatialData(path)
sce <- aggregateImage(spd)
cd <- data.frame(colData(sce), z=assay(sce)[1, ])
ggplot(cd, aes(x, y, col=z)) + geom_point() +
  scale_color_viridis_c() + scale_y_reverse()

}
\author{
Helena L. Crowell
}
