% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andalusia.R
\docType{data}
\name{andalusia_predictors}
\alias{andalusia_predictors}
\title{Predictor names for the dataset \code{andalusia}}
\format{
Character vector of length 20.
}
\usage{
data(andalusia_predictors)
}
\description{
Character vector of 20 predictor variable names corresponding to the layers of the environmental raster returned by \code{\link[=andalusia_extra]{andalusia_extra()}}, covering Landsat reflectance (7), rainfall (2), solar radiation (2), temperature (4), and topography (5). These are \strong{not} columns in \link{andalusia}; use \code{terra::extract()} to attach them to the point data.
}
\examples{
data(andalusia_predictors)
andalusia_predictors
}
\seealso{
Other andalusia:
\code{\link{andalusia}},
\code{\link[=andalusia_extra]{andalusia_extra()}},
\code{\link{andalusia_responses}}
}
\concept{andalusia}
\keyword{datasets}
