% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andalusia.R
\docType{data}
\name{andalusia_responses}
\alias{andalusia_responses}
\title{Response names for the dataset \code{andalusia}}
\format{
Character vector of length 2.
}
\usage{
data(andalusia_responses)
}
\description{
Character vector of length 2 containing the names of the
response variables in \link{andalusia}: \code{"species"} (character, species name or \code{"background"} for 90 species plus background points) and \code{"presence"} (binary integer, 1 = confirmed species presence, 0 = background point).
}
\examples{
data(andalusia_responses)
andalusia_responses
}
\seealso{
Other andalusia:
\code{\link{andalusia}},
\code{\link[=andalusia_extra]{andalusia_extra()}},
\code{\link{andalusia_predictors}}
}
\concept{andalusia}
\keyword{datasets}
