% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine two \code{SpatialData} objects}
\usage{
\S4method{combine}{SpatialData,SpatialData}(x, y, ...)
}
\arguments{
\item{x, y}{\code{SpatialData} objects to combine.}

\item{...}{ignored.}
}
\value{
For \code{centroids}, a table (\code{data.frame} or \code{matrix}) 
of spatial coordinates (if \code{as="list"}, split by instance);
for extend, a length-2 numeric list of x- and y-ranges.
}
\description{
Combine two \code{SpatialData} objects
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, anndataR=TRUE)

y <- combine(x, x)
imageNames(y)
region(table(y, 1))
region(table(y, 2))

}
