% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\alias{combine,SpatialData,SpatialData-method}
\title{Combine \code{SpatialData} objects}
\usage{
\S4method{combine}{SpatialData,SpatialData}(x, y, ...)
}
\arguments{
\item{x, y}{\code{SpatialData} objects to combine.}

\item{...}{ignored.}
}
\value{
A \code{SpatialData} objects containing all elements
from \code{x} and \code{y} with names made unique.
}
\description{
Combine \code{SpatialData} objects
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, anndataR=TRUE)

y <- combine(x, x)
imageNames(y)
region(table(y, 1))
region(table(y, 2))
}
