% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\docType{data}
\name{communities_responses}
\alias{communities_responses}
\title{Response variable names for the dataset \code{communities}}
\format{
A character vector of length 6.
}
\usage{
data(communities_responses)
}
\description{
Character vector of 6 response variable names from \link{communities}.
}
\examples{
data(communities_responses)
communities_responses
}
\seealso{
Other communities: 
\code{\link{communities}},
\code{\link{communities_extra_2010}()},
\code{\link{communities_extra_2050}()},
\code{\link{communities_extra_2100}()},
\code{\link{communities_predictors}}
}
\concept{communities}
\keyword{datasets}
