% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_utils.R
\name{coord-utils}
\alias{coord-utils}
\alias{axes}
\alias{CTname}
\alias{CTtype}
\alias{CTdata}
\alias{CTpath}
\alias{CTgraph}
\alias{addCT}
\alias{rmvCT}
\alias{axes,Zattrs-method}
\alias{axes,SpatialDataElement-method}
\alias{CTdata,Zattrs-method}
\alias{CTdata,SpatialDataElement-method}
\alias{CTtype,Zattrs-method}
\alias{CTtype,SpatialDataElement-method}
\alias{CTname,Zattrs-method}
\alias{CTname,SpatialDataElement-method}
\alias{CTname,SpatialData-method}
\alias{CTgraph,SpatialData-method}
\alias{CTgraph,SpatialDataElement-method}
\alias{CTgraph,ANY-method}
\alias{CTpath,SpatialData-method}
\alias{CTpath,SpatialDataElement-method}
\alias{rmvCT,SpatialDataElement-method}
\alias{rmvCT,Zattrs-method}
\alias{addCT,SpatialDataElement-method}
\alias{addCT,Zattrs-method}
\title{Coordinate transformations}
\usage{
\S4method{axes}{Zattrs}(x, ...)

\S4method{axes}{SpatialDataElement}(x, ...)

\S4method{CTdata}{Zattrs}(x, ...)

\S4method{CTdata}{SpatialDataElement}(x, ...)

\S4method{CTtype}{Zattrs}(x, ...)

\S4method{CTtype}{SpatialDataElement}(x, ...)

\S4method{CTname}{Zattrs}(x, ...)

\S4method{CTname}{SpatialDataElement}(x, ...)

\S4method{CTname}{SpatialData}(x, ...)

\S4method{CTgraph}{SpatialData}(x)

\S4method{CTgraph}{SpatialDataElement}(x)

\S4method{CTgraph}{ANY}(x)

\S4method{CTpath}{SpatialData}(x, i, j)

\S4method{CTpath}{SpatialDataElement}(x, j)

\S4method{rmvCT}{SpatialDataElement}(x, i)

\S4method{rmvCT}{Zattrs}(x, i)

\S4method{addCT}{SpatialDataElement}(x, name, type, data)

\S4method{addCT}{Zattrs}(x, name, type = "identity", data = NULL)
}
\arguments{
\item{x}{\code{SpatialData}, an element, or \code{Zattrs}.}

\item{...}{option arguments passed to and from other methods.}

\item{i}{for \code{CTpath}, source node label; else, string or 
scalar integer giving the name or index of a coordinate space.}

\item{j}{character string; name of target coordinate space.}

\item{name}{character(1); name of coordinate space}

\item{type}{character(1); type of transformation}

\item{data}{transformation data; size and shape depend on transformation and
element type (e.g., numeric(1) for rotation, numeric(2) for scaling in 2D)}
}
\description{
Coordinate transformations
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# element-wise
g <- CTgraph(y <- image(x))
graph::nodes(g)
CTpath(y, "global")

# object-wide
g <- CTgraph(x)
graphics.off(); graph::plot(g)

# retrieve transformation from element to target space
CTpath(x, "blobs_labels", "sequence")

# view available coordinate transformations
CTdata(z <- meta(label(x)))

# add
addCT(z, "scale", "scale", c(12, 34)) # can't overwrite
CTdata(addCT(z, "new", "translation", c(12, 34)))

# rmv
CTdata(rmvCT(z, 2)) # by index
CTdata(rmvCT(z, "scale")) # by name
CTdata(rmvCT(z, 1)) # identity is protected
}
