% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\alias{rmvCT}
\alias{rmvCT,SpatialDataElement-method}
\alias{rmvCT,Zattrs-method}
\alias{addCT,SpatialDataElement-method}
\alias{addCT,Zattrs-method}
\title{Coordinate transformations}
\usage{
\S4method{rmvCT}{SpatialDataElement}(x, i)

\S4method{rmvCT}{Zattrs}(x, i)

\S4method{addCT}{SpatialDataElement}(x, name, type, data)

\S4method{addCT}{Zattrs}(x, name, type = "identity", data = NULL)
}
\description{
Coordinate transformations
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# view available coordinate transformations
coordTransData(z <- meta(label(x)))

# add
addCT(z, "scale", "scale", c(12, 34)) # can't overwrite
coordTransData(addCT(z, "new", "translation", c(12, 34)))

# rmv
coordTransData(rmvCT(z, 2)) # by index
coordTransData(rmvCT(z, "scale")) # by name
coordTransData(rmvCT(z, 1)) # identity is protected
}
