% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr_utils.R
\name{create_zarr}
\alias{create_zarr}
\title{create_zarr}
\usage{
create_zarr(name, dir, version = "v2")
}
\arguments{
\item{name}{prefix of the zarr store, e.g. <name>.zarr}

\item{dir}{the location of zarr store, e.g. <dir>/<name>.zarr}

\item{version}{zarr version}
}
\description{
create zarr store
}
\examples{
dir.create(td <- tempfile())
zarr_name <- "test"
create_zarr(dir = td, prefix = "test")
dir.exists(file.path(td, "test.zarr"))

}
