% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata_code.R
\docType{data}
\name{demo_spatialdata}
\alias{demo_spatialdata}
\title{a data.frame with information about available resources}
\format{
An object of class \code{data.frame} with 10 rows and 5 columns.
}
\usage{
demo_spatialdata
}
\description{
a data.frame with information about available resources
}
\note{
This information was scraped from [scverse spatialdata](https://spatialdata.scverse.org/en/latest/tutorials/notebooks/datasets/README.html) download site on 5 Dec 2024.
}
\keyword{datasets}
