% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{ImageArray}
\alias{ImageArray}
\alias{dim,ImageArray-method}
\title{ImageArray}
\usage{
ImageArray(data = array(), meta = Zattrs(), metadata = list(), ...)

\S4method{dim}{ImageArray}(x)
}
\value{
...
}
\description{
ImageArray
}
\examples{
# TODO

}
