% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{ImageArray}
\alias{ImageArray}
\alias{data,ImageArray-method}
\alias{dim,ImageArray-method}
\title{The `ImageArray` class}
\usage{
ImageArray(data = list(), meta = Zattrs(), metadata = list(), ...)

\S4method{data}{ImageArray}(x, scale = 1)

\S4method{dim}{ImageArray}(x)
}
\arguments{
\item{data}{a list of arrays}

\item{meta}{...}

\item{metadata}{....}

\item{x}{\code{ImageArray}}

\item{scale}{scalar index specifying which resolution to extract.}
}
\value{
\code{ImageArray}
}
\description{
The `ImageArray` class
}
\examples{
dir.create(td <- tempfile())
pa <- unzip_merfish_demo(td)
pa <- file.path(pa, "images", "rasterized")
(ia <- readImage(pa))

}
