% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/ImageArray.R
\name{metadata,ImageArray-method}
\alias{metadata,ImageArray-method}
\alias{dim,ImageArray-method}
\alias{dimnames,ImageArray-method}
\alias{[,ImageArray,ANY,ANY,ANY-method}
\alias{getArrayElement,ImageArray-method}
\alias{as.array,ImageArray-method}
\alias{aperm,ImageArray-method}
\alias{ImageArray}
\title{#' @rdname ImageArray
#' @export
setMethod("extract_array", "ImageArray", function(x, index) {
  extract_array(data(x), index)
})}
\usage{
\S4method{metadata}{ImageArray}(x)

\S4method{dim}{ImageArray}(x)

\S4method{dimnames}{ImageArray}(x)

\S4method{[}{ImageArray,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{getArrayElement}{ImageArray}(x, subscripts)

\S4method{as.array}{ImageArray}(x)

\S4method{aperm}{ImageArray}(a, perm)

ImageArray(data, metadata = list(), ...)
}
\arguments{
\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{A \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}
}
\description{
...
}
\examples{
dir <- "extdata/mibitof/images/point8_image"
zarr <- system.file(file.path(dir, "0"), package = "SpatialData")
json <- system.file(file.path(dir, ".zattrs"), package = "SpatialData")

library(Rarr)
library(jsonlite)

za <- read_zarr_array(zarr)
md <- fromJSON(json)
(ia <- ImageArray(za, md))

}
