% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{ImageArray}
\alias{ImageArray}
\alias{dim,ImageArray-method}
\title{The `ImageArray` class}
\usage{
ImageArray(data_list = list(), meta = Zattrs(), metadata = list(), ...)

\S4method{dim}{ImageArray}(x)
}
\arguments{
\item{data_list}{a list of arrays}

\item{meta}{...}

\item{metadata}{....}
}
\value{
\code{ImageArray}
}
\description{
The `ImageArray` class
}
\examples{
dir.create(td <- tempfile())
pa <- unzip_merfish_demo(td)
pa <- file.path(pa, "images", "rasterized")
(ia <- readImage(pa))

a <- as.array(data(ia))
a <- aperm(a, c(2,3,1))
plot(EBImage::Image(a/255))

}
