% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageArray.R
\name{ImageArray}
\alias{ImageArray}
\alias{channels,ImageArray-method}
\alias{channels}
\alias{channels,ANY-method}
\alias{data,ImageArray-method}
\alias{dim,ImageArray-method}
\alias{length,ImageArray-method}
\alias{[,ImageArray,ANY,ANY,ANY-method}
\alias{getZarrArrayPath}
\title{The `ImageArray` class}
\usage{
ImageArray(data = list(), meta = Zattrs(), metadata = list(), ...)

\S4method{channels}{ImageArray}(x, ...)

\S4method{channels}{ANY}(x, ...)

\S4method{data}{ImageArray}(x, k = 1)

\S4method{dim}{ImageArray}(x)

\S4method{length}{ImageArray}(x)

\S4method{[}{ImageArray,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)

getZarrArrayPath(x)
}
\arguments{
\item{data}{list of \code{\link[Rarr]{ZarrArray}}s}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{ImageArray}}

\item{k}{scalar index specifying which scale to extract.}

\item{i, j}{indices specifying elements to extract.}

\item{drop}{ignored.}
}
\value{
\code{ImageArray}
}
\description{
The `ImageArray` class
}
\examples{
dir.create(td <- tempfile())
pa <- unzip_merfish_demo(td)
pa <- file.path(pa, "images", "rasterized")
(ia <- readImage(pa))

}
