% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/ImageArray.R, R/readPoints.R,
%   R/transformations.R
\name{metadata,ImageArray-method}
\alias{metadata,ImageArray-method}
\alias{dim,ImageArray-method}
\alias{dimnames,ImageArray-method}
\alias{[,ImageArray,ANY,ANY,ANY-method}
\alias{getArrayElement,ImageArray-method}
\alias{as.array,ImageArray-method}
\alias{aperm,ImageArray-method}
\alias{ImageArray}
\alias{translateImage}
\alias{scaleImage}
\alias{rotateImage}
\alias{readPoints}
\alias{coords,ImageArray-method}
\alias{scaleImage,ImageArray-method}
\alias{rotateImage,ImageArray-method}
\alias{translateImage,ImageArray-method}
\title{#' @rdname ImageArray
#' @export
setMethod("extract_array", "ImageArray", function(x, index) {
  extract_array(x@data, index)
})}
\usage{
\S4method{metadata}{ImageArray}(x)

\S4method{dim}{ImageArray}(x)

\S4method{dimnames}{ImageArray}(x)

\S4method{[}{ImageArray,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{getArrayElement}{ImageArray}(x, subscripts)

\S4method{as.array}{ImageArray}(x)

\S4method{aperm}{ImageArray}(a, perm)

ImageArray(data = array(), metadata = list(), ...)

readPoints(path, ...)

\S4method{coords}{ImageArray}(x)

\S4method{scaleImage}{ImageArray}(x, t = rep(1, length(dim(x))))

\S4method{rotateImage}{ImageArray}(x, t = 0)

\S4method{translateImage}{ImageArray}(x, t = 0)
}
\arguments{
\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{An \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}

\item{path}{A character string specifying
a .parquet file-containing directory.}
}
\description{
...

...
}
\section{Transformations}{

\describe{
\item{\code{translateImage}}{
  translates xy coordinates according to \code{t}
  (see \code{\link[EBImage:resize]{translate}})}
\item{\code{scaleImage}}{
  scales the image to the desired dimensions
  (see \code{\link[EBImage:resize]{resize}})}
\item{\code{rotateImage}}{
  rotates the image clockwise by
  the given angle around the origin.
  (see \code{\link[EBImage:resize]{rotate}})}
}
}

\examples{
dir <- "extdata/mibitof/images/point8_image"
zarr <- system.file(file.path(dir, "0"), package = "SpatialData")
json <- system.file(file.path(dir, ".zattrs"), package = "SpatialData")

library(Rarr)
library(jsonlite)

za <- read_zarr_array(zarr)
md <- fromJSON(json)
(ia <- ImageArray(za, md))

path <- "extdata/blobs/points/blobs_points"
path <- system.file(path, package = "SpatialData")
(df <- readPoints(path))

}
