% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\docType{data}
\name{interaction_responses}
\alias{interaction_responses}
\title{Response variable names for the dataset \code{interaction}}
\format{
A character vector of length 3.
}
\usage{
data(interaction_responses)
}
\description{
Character vector of 3 response variable names from \link{interaction}.
}
\examples{
data(interaction_responses)
interaction_responses
}
\seealso{
Other interaction:
\code{\link{interaction}},
\code{\link[=interaction_extra]{interaction_extra()}},
\code{\link{interaction_predictors}}
}
\concept{interaction}
\keyword{datasets}
