% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelArray.R
\name{LabelArray}
\alias{LabelArray}
\alias{dim,LabelArray-method}
\title{LabelArray}
\usage{
LabelArray(data = array(), meta = Zattrs(), metadata = list(), ...)

\S4method{dim}{LabelArray}(x)
}
\value{
...
}
\description{
LabelArray
}
\examples{
# TODO

}
