% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelArray.R
\name{LabelArray}
\alias{LabelArray}
\alias{dim,LabelArray-method}
\title{The `LabelArray` class}
\usage{
LabelArray(data = array(), meta = Zattrs(), metadata = list(), ...)

\S4method{dim}{LabelArray}(x)
}
\arguments{
\item{data}{list of \code{\link[Rarr]{ZarrArray}}s}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{LabelArray}}
}
\value{
\code{LabelArray}
}
\description{
The `LabelArray` class (LA) is defined to store labels in `array` or
`data.frame` format into the dedicated `@data` slot together with the `meta`
attributes defined in the Zattr files.
Optional metadata can be stored in the `metadata` list.

Defined methods are:
- dim: returns the dimensions of the `@data` slot.
}
\examples{
y <- file.path("extdata", "blobs.zarr")
y <- file.path(system.file(y, package="SpatialData"), "labels",
    "blobs_labels")
(l <- readLabel(y))

}
