% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelArray.R
\name{LabelArray}
\alias{LabelArray}
\title{The \code{LabelArray} class}
\usage{
LabelArray(data = array(), meta = Zattrs(), metadata = list(), ...)
}
\arguments{
\item{data}{list of \code{\link[Rarr]{ZarrArray}}s}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{LabelArray}}

\item{i, j}{indices specifying elements to extract.}

\item{drop}{ignored.}
}
\value{
\code{LabelArray}
}
\description{
The \code{LabelArray} class stores \code{SpatialData} elements from its 
\code{"labels"} layers. These are represented as a \code{ZarrMatrix} 
(\code{data} slot) associated with .zattrs stored as \code{\link{Zattrs}} 
(\code{meta} slot); a list of \code{metadata} stores other arbitrary info.

Currently defined methods (here, \code{x} is a \code{LabelArray}):
\itemize{
\item \code{data/meta(x)} to access underlying \code{ZarrMatrix/Zattrs}
\item \code{dim(x)} returns the dimensions of \code{data(x)}
}
}
\examples{
# TODO

}
