% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linaria.R
\name{linaria_extra}
\alias{linaria_extra}
\title{Download Environmental Raster for linaria}
\usage{
linaria_extra()
}
\value{
SpatRaster object with 20 layers.
}
\description{
Downloads and reads the 20-band environmental raster associated with the \link{linaria} dataset
from the \href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository. Writes the file \code{linaria_env.tif} in the working directory and returns it as a \code{spatRaster} object.
}
\seealso{
Other linaria: 
\code{\link{linaria}},
\code{\link{linaria_predictors}},
\code{\link{linaria_responses}}
}
\concept{linaria}
