% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\alias{mask,SpatialData,ANY,ANY-method}
\title{Masking}
\usage{
\S4method{mask}{SpatialData,ANY,ANY}(x, i, j, how = NULL)
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i, j}{character string; names of elements to mask,
specifically, \code{i} will be masked by \code{j},
adding a \code{table} for \code{j} in \code{x}.}

\item{how}{character string; statistic to use for masking.
Defaults to "mean" when masking images, ignored when masking points.}

\item{...}{optional arguments passed to and from other methods.}
}
\value{
Input \code{SpatialData} object \code{x} with an additional table named 
\code{<j>_masking_<i>}; or a \code{SingleCellExperiment} object when 
masking elements directly (i.e., without \code{x} as input).
}
\description{
Masking operations serve to aggregate data across layers, e.g., 
counting points in shapes, averaging image channels by labels, etc.
For added flexibility, these may be carried out directly between elements,
or using an input \code{SpatialData} object and specifying element names.
}
\examples{
library(SingleCellExperiment)
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# count points in shapes
y <- mask(x, "blobs_points", "blobs_circles")
(sce <- mask(i=point(x), j=shape(x, 1)))
identical(assay(table(y)), assay(sce))

# average image channels by labels
y <- mask(x, "blobs_image", "blobs_labels")
(sce <- mask(i=image(x), j=label(x)))
identical(assay(table(y)), assay(sce))

}
