% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\alias{mask,ImageArray,LabelArray-method}
\alias{mask,ImageArray,ShapeFrame-method}
\alias{mask,LabelArray,ShapeFrame-method}
\title{Mask an element by another}
\usage{
\S4method{mask}{ImageArray,LabelArray}(x, y, fun = mean, name = "X")

\S4method{mask}{ImageArray,ShapeFrame}(x, y, fun = mean, name = "X")

\S4method{mask}{LabelArray,ShapeFrame}(x, y, fun = mean, name = "X")
}
\arguments{
\item{x, y}{A \code{\link{SpatialData}} element
(see Usage for currently supported methods).}

\item{fun}{Function to use for aggregation.}

\item{name}{Character string specifying
the name of the output SCE's `assay`.}
}
\value{
A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
where rows/columns correspond to data from \code{x/y}.
}
\description{
...
}
\examples{
dir <- file.path("extdata", "blobs")
dir <- system.file(dir, package="SpatialData")
spd <- readSpatialData(dir)
(sce <- mask(image(spd), label(spd)))

}
\author{
Helena L. Crowell
}
