% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\alias{mask,SpatialData-method}
\title{Masking}
\usage{
\S4method{mask}{SpatialData}(x, i, j)
}
\arguments{
\item{x, y}{\code{\link{SpatialData}} element}
}
\value{
\code{\link{SingleCellExperiment}}
}
\description{
...
}
\examples{
library(SingleCellExperiment)
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# count points in circles
x <- mask(x, "blobs_points", "blobs_circles")
(t <- table(x))

}
