% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{show,SpatialData-method}
\alias{show,sdArray-method}
\alias{show,PointFrame-method}
\alias{show,ShapeFrame-method}
\title{Miscellaneous `SpatialData` methods}
\usage{
\S4method{show}{SpatialData}(object)

\S4method{show}{sdArray}(object)

\S4method{show}{PointFrame}(object)

\S4method{show}{ShapeFrame}(object)
}
\arguments{
\item{object}{\code{\link{SpatialData}} object or one of its elements,
i.e., an \code{Image/LabelArray} or \code{Point/ShapeFrame}.}
}
\value{
\code{NULL}
}
\description{
Miscellaneous methods (e.g., \code{show}) for the
\code{\link{SpatialData}} class and its elements.
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")
(sd <- readSpatialData(zs))

# show element
image(sd)
label(sd)
point(sd)
shape(sd)

# show .zattrs
meta(label(sd))
meta(image(sd, 2))
}
