% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{show,SpatialData-method}
\alias{show,ImageArray-method}
\alias{show,LabelArray-method}
\alias{show,PointFrame-method}
\alias{show,ShapeFrame-method}
\title{Miscellaneous `Miro` methods}
\usage{
\S4method{show}{SpatialData}(object)

\S4method{show}{ImageArray}(object)

\S4method{show}{LabelArray}(object)

\S4method{show}{PointFrame}(object)

\S4method{show}{ShapeFrame}(object)
}
\value{
\code{NULL}
}
\description{
...
}
\examples{
# TODO
}
\author{
Helena L. Crowell
}
