% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\docType{data}
\name{neanderthal}
\alias{neanderthal}
\title{Neanderthal Presence and Pseudo-Absence Records with Environmental Predictors}
\format{
sf data.frame
}
\usage{
data(neanderthal)
}
\description{
\code{sf} dataframe with \code{POINT} geometry containing 245 records of Neanderthal presence and pseudo-absence sites from Marine Isotope Stage 5e (Last Interglacial) in Europe and the Near East, 1 response variable (see \link{neanderthal_response}), and 25 predictors (see \link{neanderthal_predictors}). Use \code{\link[=neanderthal_extra]{neanderthal_extra()}} to download the associated environmental raster.

\strong{Response}

The response variable \code{presence} is a binary integer:
\itemize{
\item \code{1}: Neanderthal presence site.
\item \code{0}: Pseudo-absence site.
}

\strong{Predictors}

\itemize{
\item \code{bio1}, \code{bio2}, ..., \code{bio19}: Palaeoclimatic variables from a Last Interglacial GCM simulation (19 variables).
\item \code{topo_aspect}, \code{topo_diversity_local}, \code{topo_diversity}, \code{topo_elev}, \code{topo_slope}, \code{topo_wetness}: SRTM-derived topographic variables.
}

\strong{Geometry:}
\itemize{
\item \code{geometry}: Point geometry (WGS84, EPSG:4326).
}

\strong{Data sources}
\itemize{
\item \strong{Presence data}: Benito, B.M., Svenning, J.-C., Kellberg-Nielsen, T., Riede, F., Gil-Romera, G., Mailund, T., Kjaergaard, P.C. and Sandel, B.S. (2017), The ecological niche and distribution of Neanderthals during the Last Interglacial. J. Biogeogr., 44: 51-61. \doi{10.1111/jbi.12845}
\item \strong{Palaeoclimatic variables (GCM simulation)}: Otto-Bliesner, B.L., Marshall, S.J., Overpeck, J.T., Miller, G.H. & Hu, A. (2006) Simulating arctic climate warmth and icefield retreat in the last interglaciation. Science, 311, 1751-1753.
\item \strong{Palaeoclimatic variables (interpolation)}: Hijmans, R.J., Cameron, S.E., Parra, J.L., Jones, P.G. & Jarvis, A. (2005) Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology, 25, 1965-1978.
\item \strong{Elevation and topography}: Jarvis A., H.I. Reuter, A. Nelson, E. Guevara, 2008. Hole-filled seamless SRTM data V4, International Centre for Tropical Agriculture (CIAT). \url{https://srtm.csi.cgiar.org}.
}
}
\examples{
data(neanderthal)
head(neanderthal)
}
\seealso{
Other neanderthal: 
\code{\link{neanderthal_extra}()},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}}
}
\concept{neanderthal}
\keyword{datasets}
