% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\name{neanderthal_extra}
\alias{neanderthal_extra}
\title{Download Environmental Raster for neanderthal}
\usage{
neanderthal_extra(dir = ".", quiet = TRUE)
}
\arguments{
\item{dir}{(optional, character) Directory to save the file. Defaults to
the current working directory.}

\item{quiet}{(optional, logical) If \code{TRUE} (default), suppresses
download messages.}
}
\value{
SpatRaster object.
}
\description{
Downloads and reads the environmental raster associated with the \link{neanderthal}
dataset from the
\href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository.
Requires the \pkg{terra} package.
}
\examples{
\dontrun{
neanderthal_env <- neanderthal_extra()
neanderthal_env
}
}
\seealso{
Other example_data: 
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}},
\code{\link{quercus}},
\code{\link{quercus_extra}()},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}},
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}},
\code{\link{vi_responses}}
}
\concept{example_data}
