% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\docType{data}
\name{neanderthal_predictors}
\alias{neanderthal_predictors}
\title{Predictor Variable Names for \code{neanderthal}}
\format{
Character vector of length 25.
}
\usage{
data(neanderthal_predictors)
}
\description{
Character vector with the names of the 25 predictor variables in \link{neanderthal}: 19 bioclimatic and 6 topographic variables.
}
\seealso{
Other example_data: 
\code{\link{neanderthal}},
\code{\link{neanderthal_extra}()},
\code{\link{neanderthal_response}},
\code{\link{quercus}},
\code{\link{quercus_extra}()},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}},
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}},
\code{\link{vi_responses}}
}
\concept{example_data}
\keyword{datasets}
