% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\docType{data}
\name{neanderthal_predictors}
\alias{neanderthal_predictors}
\title{Predictor variable names for Neanderthal distribution examples}
\format{
A character vector of length 25.
}
\usage{
data(neanderthal_predictors)
}
\description{
Character vector of predictor variable names from \code{neanderthal} (19 bioclimatic and 6 topographic variables).
}
\seealso{
Other data: 
\code{\link{drylands}},
\code{\link{drylands_predictors}},
\code{\link{drylands_responses}},
\code{\link{neanderthal}},
\code{\link{neanderthal_response}},
\code{\link{plantae}},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}},
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}}
}
\concept{data}
\keyword{datasets}
