% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\docType{data}
\name{neanderthal_predictors}
\alias{neanderthal_predictors}
\title{Predictor variable names for the dataset \code{neanderthal}}
\format{
A character vector of length 25.
}
\usage{
data(neanderthal_predictors)
}
\description{
Character vector of 25 predictor variable names from \link{neanderthal}.
}
\examples{
data(neanderthal_predictors)
neanderthal_predictors
}
\seealso{
Other neanderthal:
\code{\link{neanderthal}},
\code{\link[=neanderthal_extra]{neanderthal_extra()}},
\code{\link{neanderthal_response}}
}
\concept{neanderthal}
\keyword{datasets}
