% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\docType{data}
\name{neanderthal_response}
\alias{neanderthal_response}
\title{Response variable name for Neanderthal distribution examples}
\format{
A character string of length 1.
}
\usage{
data(neanderthal_response)
}
\description{
Character string containing the name of the response variable in \code{neanderthal}: "presence".
}
\seealso{
Other data: 
\code{\link{drylands}},
\code{\link{drylands_predictors}},
\code{\link{drylands_responses}},
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{plantae}},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}},
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}}
}
\concept{data}
\keyword{datasets}
