% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neanderthal_response}
\alias{neanderthal_response}
\title{Response variable name for Neanderthal distribution examples}
\format{
A character string of length 1.
}
\usage{
data(neanderthal_response)
}
\description{
Character string containing the name of the response variable in \code{neanderthal}: "presence".
}
\seealso{
Other data: 
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{plants}},
\code{\link{plants_predictors}},
\code{\link{plants_response}},
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}}
}
\concept{data}
\keyword{datasets}
