% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\docType{data}
\name{neanderthal_response}
\alias{neanderthal_response}
\title{Response Variable Name for \code{neanderthal}}
\format{
Character string of length 1.
}
\usage{
data(neanderthal_response)
}
\description{
Character string with the name of the response variable in \link{neanderthal}: \code{"presence"}.
}
\seealso{
Other example_data: 
\code{\link{neanderthal}},
\code{\link{neanderthal_extra}()},
\code{\link{neanderthal_predictors}},
\code{\link{quercus}},
\code{\link{quercus_extra}()},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}},
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}},
\code{\link{vi_responses}}
}
\concept{example_data}
\keyword{datasets}
