% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\docType{data}
\name{neanderthal_response}
\alias{neanderthal_response}
\title{Response variable name for the dataset \code{neanderthal}}
\format{
A character string of length 1.
}
\usage{
data(neanderthal_response)
}
\description{
Character string with the name of the response variable in \link{neanderthal}.
}
\examples{
data(neanderthal_response)
neanderthal_response
}
\seealso{
Other neanderthal:
\code{\link{neanderthal}},
\code{\link[=neanderthal_extra]{neanderthal_extra()}},
\code{\link{neanderthal_predictors}}
}
\concept{neanderthal}
\keyword{datasets}
