% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{path_to_10x_xen_demo}
\alias{path_to_10x_xen_demo}
\title{provide path to a zip file from 10x genomics for Xenium platform}
\usage{
path_to_10x_xen_demo(zipname = "Xenium_V1_human_Breast_2fov_outs.zip")
}
\arguments{
\item{zipname}{character(1) name of zip archive to find}
}
\description{
provide path to a zip file from 10x genomics for Xenium platform
}
\examples{
path_to_10x_xen_demo()
# see ?use_sdio
}
