% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\docType{data}
\name{plantae_east}
\alias{plantae_east}
\title{Eastern Hemisphere subset of \code{plantae}}
\format{
\code{sf} dataframe.
}
\usage{
data(plantae_east)
}
\description{
Subset of the \link{plantae} dataset filtered to American ecoregions (\code{ecoregion_continent != "Americas"}) with \code{richness_species} (overall plant species richness) as the only response variable. All 84 predictor variables and identifier columns in \link{plantae} are retained.
}
\seealso{
Other plantae: 
\code{\link{plantae}},
\code{\link{plantae_extra}()},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}},
\code{\link{plantae_west}}
}
\concept{plantae}
\keyword{datasets}
