% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\docType{data}
\name{plantae_east}
\alias{plantae_east}
\title{Eastern Hemisphere subset of \code{plantae}}
\format{
An sf data frame with 434 rows and 91 columns.
}
\usage{
data(plantae_east)
}
\description{
Subset of the \link{plantae} dataset filtered to non-American ecoregions (\code{ecoregion_continent != "Americas"}) with \code{richness_species} (overall plant species richness) as the only response variable. All 84 predictor variables and identifier columns in \link{plantae} are retained.
}
\examples{
data(plantae_east)
colnames(plantae_east)
nrow(plantae_east)
ncol(plantae_east)
}
\seealso{
Other plantae:
\code{\link{plantae}},
\code{\link[=plantae_extra]{plantae_extra()}},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}},
\code{\link{plantae_west}}
}
\concept{plantae}
\keyword{datasets}
