% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\docType{data}
\name{plantae_predictors}
\alias{plantae_predictors}
\title{Predictor variable names for plantae dataset}
\format{
A character vector of length 96.
}
\usage{
data(plantae_predictors)
}
\description{
Character vector of 96 predictor variable names from \code{plantae}, covering
sampling bias (7), geographic/geometric (14), human impact (5), climate change (2),
landcover (3), fragmentation (19), atmospheric (13), precipitation (8), temperature (11),
soil properties (6), soil temperature (4), and NDVI (4).
}
\seealso{
Other plantae: 
\code{\link{plantae}},
\code{\link{plantae_responses}}
}
\concept{plantae}
\keyword{datasets}
