% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\docType{data}
\name{plantae_predictors}
\alias{plantae_predictors}
\title{Predictor variable names for the dataset \code{plantae}}
\format{
A character vector of length 84.
}
\usage{
data(plantae_predictors)
}
\description{
Character vector of 84 predictor variable names from \code{plantae}.
}
\examples{
data(plantae_predictors)
plantae_predictors
}
\seealso{
Other plantae: 
\code{\link{plantae}},
\code{\link{plantae_east}},
\code{\link{plantae_extra}()},
\code{\link{plantae_responses}},
\code{\link{plantae_west}}
}
\concept{plantae}
\keyword{datasets}
