% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantae.R
\docType{data}
\name{plantae_west}
\alias{plantae_west}
\title{Western Hemisphere subset of \code{plantae}}
\format{
An sf data frame with 228 rows and 91 columns.
}
\usage{
data(plantae_west)
}
\description{
Subset of the \link{plantae} dataset filtered to American ecoregions (\code{ecoregion_continent == "Americas"}) with \code{richness_species} (overall plant species richness) as the only response variable. All 84 predictor variables and identifier columns in \link{plantae} are retained.
}
\examples{
data(plantae_west)
colnames(plantae_west)
nrow(plantae_west)
ncol(plantae_west)
}
\seealso{
Other plantae: 
\code{\link{plantae}},
\code{\link{plantae_east}},
\code{\link{plantae_extra}()},
\code{\link{plantae_predictors}},
\code{\link{plantae_responses}}
}
\concept{plantae}
\keyword{datasets}
