% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants_predictors}
\alias{plants_predictors}
\title{Predictor variable names for plant richness examples}
\format{
A character vector of length 17.
}
\usage{
data(plants_predictors)
}
\description{
Character vector of predictor variable names from \code{plants} (columns 3 to 19).
}
\seealso{
Other data: 
\code{\link{drylands}},
\code{\link{drylands_predictors}},
\code{\link{drylands_responses}},
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}},
\code{\link{plants}},
\code{\link{plants_response}},
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}}
}
\concept{data}
\keyword{datasets}
