% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_nodes.R
\name{plotCoordGraph}
\alias{plotCoordGraph}
\title{given a graphNEL instance, nodes with nchar > max are split and hyphenated at
character position floor(nchar/fac)}
\usage{
plotCoordGraph(g, cex = 0.6)
}
\description{
given a graphNEL instance, nodes with nchar > max are split and hyphenated at
character position floor(nchar/fac)
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)
g <- SpatialData:::.coord2graph(x)
plotCoordGraph(g, cex=.6)
}
