% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpatialData}
\alias{plotSpatialData}
\alias{plotImage}
\alias{plotLabel}
\alias{plotPoint}
\alias{plotShape}
\alias{plotLabel,SpatialData-method}
\title{`SpatialData` visualization}
\usage{
plotSpatialData()

\S4method{plotLabel}{SpatialData}(x, i = 1, c = NULL, a = 0.5, pal = hcl.colors(11, "Spectral"))
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i}{element to use from a given layer.}

\item{c, f, s, a}{plotting aesthetics; color, fill, size, alpha.}

\item{pal}{character vector of colors; will interpolate 
automatically when insufficient values are provided.}

\item{j}{name of target coordinate system.}

\item{k}{index of the scale of an image (default: NULL)}
}
\value{
ggplot
}
\description{
...
}
\examples{
tf <- tempfile()
dir.create(tf)
base <- unzip_merfish_demo(tf)
(x <- readSpatialData(base, tables=FALSE))

plotImage(x)

}
