% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\alias{plotSD}
\title{Plot `SpatialData` elements}
\usage{
plotSD(
  x,
  image = 1,
  label = 1,
  shape = 1,
  alpha.label = 1/3,
  alpha.shape = 1,
  ...
)
}
\arguments{
\item{x}{An \code{ImageArray} object.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
\code{NULL}
}
\description{
...
}
\examples{
path <- system.file("extdata", "raccoon", package="SpatialData")
plotSD(readSpatialData(path))

path <- system.file("extdata", "blobs", package="SpatialData")
plotSD(readSpatialData(path))
}
\author{
Helena L. Crowell
}
