% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\alias{plotSD}
\alias{plotElement}
\alias{plotSD,SpatialData}
\alias{plotElement,ImageArray}
\alias{plotElement,LabelArray}
\alias{plotElement,ShapeFrame}
\alias{plotElement,ImageArray-method}
\alias{plotElement,LabelArray-method}
\alias{plotElement,ShapeFrame-method}
\title{Plot `SpatialData` elements}
\usage{
\S4method{plotElement}{ImageArray}(x, coord = NULL, ...)

\S4method{plotElement}{LabelArray}(x, coord = NULL, ...)

\S4method{plotElement}{ShapeFrame}(x, coord = NULL, ...)

plotSD(
  x,
  image,
  label,
  shape,
  alpha.label = 1/3,
  alpha.shape = 1,
  fill.shape = "lightgrey",
  col.shape = NA,
  coord = NULL
)
}
\arguments{
\item{x}{A \code{\link{SpatialData}} object.}

\item{coord}{A character string specifying the target coordinate system.
If \code{NULL}, defaults to the first available shared coordinates.}

\item{image, label, shape}{A scalar integer (index) or character string (identifier)
specifying the entity to include (can be \code{NULL}).}

\item{alpha.label, alpha.shape, fill.shape, col.shape}{Plotting aesthetics.}
}
\value{
\code{ggplot2}
}
\description{
...
}
\examples{
path <- system.file("extdata", "raccoon", package="SpatialData")
sd <- readSpatialData(path)

# by element
plotElement(image(sd))
plotElement(label(sd))
plotElement(shape(sd))

# layered
plotSD(sd)
plotSD(sd, image=NULL)
plotSD(sd, label=NULL, fill.shape="pink")
plotSD(sd, shape=NULL, alpha.label=0.2)
}
\author{
Helena L. Crowell
}
