% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R, R/plotLabel.R, R/plotPoint.R,
%   R/plotShape.R, R/plotting.R
\name{plotImage,SpatialData-method}
\alias{plotImage,SpatialData-method}
\alias{plotLabel,SpatialData-method}
\alias{plotPoint,SpatialData-method}
\alias{plotShape,SpatialData-method}
\alias{plotting}
\alias{plotImage}
\alias{plotLabel}
\alias{plotShape}
\alias{plotPoint}
\title{`SpatialData` plotting}
\usage{
\S4method{plotImage}{SpatialData}(x, ..., i = 1, coord = 1, width = 400)

\S4method{plotLabel}{SpatialData}(x, ..., i = 1, coord = 1, fill = NULL, alpha = 1/3, width = 400)

\S4method{plotPoint}{SpatialData}(x, ..., i = 1, coord = NULL)

\S4method{plotShape}{SpatialData}(x, ..., i = 1, coord = NULL)
}
\arguments{
\item{x}{A \code{\link{SpatialData}} object.}

\item{...}{Additional graphical parameters passed to
\code{\link[ggplot2]{geom_polygon}} for shapes;
for labels, an \code{alpha} value can be passed;
ignored otherwise.}

\item{i}{Scalar integer or character string specifying which element to plot.}

\item{coord}{A character string specifying the target coordinate system.
If \code{NULL}, defaults to the first available shared coordinates.}

\item{width}{Scalar integer specifying the output plot's width (in pixel).}
}
\value{
\code{ggplot2}
}
\description{
...
}
\examples{
path <- system.file("extdata", "raccoon", package="SpatialData")
spd <- readSpatialData(path)

# by element
i <- plotImage(spd)
l <- plotLabel(spd)
s <- plotShape(spd)

# putting it together
i+l+s
}
\author{
Helena L. Crowell
}
