% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointFrame.R
\name{PointFrame}
\alias{PointFrame}
\alias{names,PointFrame-method}
\alias{dim,PointFrame-method}
\alias{length,PointFrame-method}
\alias{$,PointFrame-method}
\alias{[[,PointFrame,ANY,ANY-method}
\alias{[,PointFrame,numeric,ANY,ANY-method}
\alias{as.data.frame,PointFrame-method}
\title{The `PointFrame` class}
\usage{
PointFrame(data = data.frame(), meta = Zattrs(), metadata = list(), ...)

\S4method{names}{PointFrame}(x)

\S4method{dim}{PointFrame}(x)

\S4method{length}{PointFrame}(x)

\S4method{$}{PointFrame}(x, name)

\S4method{[[}{PointFrame,ANY,ANY}(x, i, j, ...)

\S4method{[}{PointFrame,numeric,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{as.data.frame}{PointFrame}(x)
}
\value{
\code{PointFrame}
}
\description{
The `PointFrame` class
}
\examples{
tf <- tempfile()
dir.create(tf)
base <- unzip_merfish_demo(tf)
x <- file.path(base, "points", "single_molecule")
(p <- readPoint(x))

head(as.data.frame(data(p)))
(q <- dplyr::filter(p, cell_type == "VISp_wm"))
plotPoint(q, c="x", s=0.2)

}
