% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quercus.R
\name{quercus_extra}
\alias{quercus_extra}
\title{Download Environmental Raster for quercus}
\usage{
quercus_extra(dir = ".", quiet = TRUE)
}
\arguments{
\item{dir}{(optional, character) Directory to save the file. Defaults to
the current working directory.}

\item{quiet}{(optional, logical) If \code{TRUE} (default), suppresses
download messages.}
}
\value{
SpatRaster object.
}
\description{
Downloads and reads the environmental raster associated with the \link{quercus}
dataset from the
\href{https://github.com/BlasBenito/spatialDataExtra}{spatialDataExtra} repository.
Requires the \pkg{terra} package.
}
\examples{
\dontrun{
quercus_env <- quercus_extra()
quercus_env
}
}
\seealso{
Other example_data: 
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{quercus_response}},
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}},
\code{\link{vi_responses}}
}
\concept{example_data}
