% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quercus.R
\docType{data}
\name{quercus_predictors}
\alias{quercus_predictors}
\title{Predictor Variable Names for \code{quercus}}
\format{
Character vector of length 31.
}
\usage{
data(quercus_predictors)
}
\description{
Character vector with the names of the 31 predictor variables in \link{quercus}: 17 bioclimatic, 4 NDVI, 4 solar radiation, 3 land cover, 2 topographic, and 1 human footprint variable.
}
\seealso{
Other example_data: 
\code{\link{neanderthal}},
\code{\link{neanderthal_extra}()},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}},
\code{\link{quercus}},
\code{\link{quercus_extra}()},
\code{\link{quercus_response}},
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}},
\code{\link{vi_responses}}
}
\concept{example_data}
\keyword{datasets}
