% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quercus_predictors}
\alias{quercus_predictors}
\title{Predictor variable names for Quercus distribution examples}
\format{
A character vector of length 31.
}
\usage{
data(quercus_predictors)
}
\description{
Character vector of predictor variable names from \code{quercus} (17 bioclimatic, 2 topographic, 3 landcover, 4 NDVI, 4 solar radiation, and 1 human footprint variable).
}
\seealso{
Other data: 
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}},
\code{\link{plants}},
\code{\link{plants_predictors}},
\code{\link{plants_response}},
\code{\link{quercus}},
\code{\link{quercus_response}}
}
\concept{data}
\keyword{datasets}
