% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quercus.R
\docType{data}
\name{quercus_predictors}
\alias{quercus_predictors}
\title{Predictor variable names for for the dataset \code{quercus}}
\format{
A character vector of length 31.
}
\usage{
data(quercus_predictors)
}
\description{
Character vector of 31 predictor variable names from \link{quercus}.
}
\examples{
data(quercus_predictors)
quercus_predictors
}
\seealso{
Other quercus:
\code{\link{quercus}},
\code{\link[=quercus_extra]{quercus_extra()}},
\code{\link{quercus_response}}
}
\concept{quercus}
\keyword{datasets}
