% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quercus.R
\docType{data}
\name{quercus_response}
\alias{quercus_response}
\title{Response Variable Name for \code{quercus}}
\format{
Character string of length 1.
}
\usage{
data(quercus_response)
}
\description{
Character string with the name of the response variable in \link{quercus}: \code{"species"}.
}
\seealso{
Other example_data: 
\code{\link{quercus}},
\code{\link{quercus_predictors}},
\code{\link{vi}},
\code{\link{vi_predictors}},
\code{\link{vi_responses}}
}
\concept{example_data}
\keyword{datasets}
