% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quercus_response}
\alias{quercus_response}
\title{Response variable name for Quercus distribution examples}
\format{
A character string of length 1.
}
\usage{
data(quercus_response)
}
\description{
Character string containing the name of the response variable in \code{quercus}: "species".
}
\seealso{
Other data: 
\code{\link{drylands}},
\code{\link{drylands_predictors}},
\code{\link{drylands_responses}},
\code{\link{neanderthal}},
\code{\link{neanderthal_predictors}},
\code{\link{neanderthal_response}},
\code{\link{quercus}},
\code{\link{quercus_predictors}}
}
\concept{data}
\keyword{datasets}
