% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query,SpatialData-method}
\title{queries}
\usage{
\S4method{query}{SpatialData}(x, i, ...)
}
\arguments{
\item{x}{\code{SpatialData} object.}

\item{i}{index or name of table to query.}

\item{...}{logic passed to \code{dplyr::filter}.}
}
\value{
\code{SpatialData} object
}
\description{
\code{query} provides a interface for table-based 
subsetting of \code{SpatialData} objects. It filters a specified 
table using \code{dplyr::filter} logic and propagates the result 
to all associated spatial elements (i.e., only instances 
present in the filtered table are kept).

For spatial cropping, see \code{\link{crop}}.
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")
sd <- readSpatialData(zs)

# filter by 'region' and propagate to shapes/points
t <- table(sd)
query(sd, i=1, region == region(t))
}
