% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query,SpatialData-method}
\alias{query,ImageArray-method}
\alias{query,LabelArray-method}
\alias{query,ShapeFrame-method}
\alias{query,PointFrame-method}
\title{spatial queries}
\usage{
\S4method{query}{SpatialData}(x, ..., i)

\S4method{query}{ImageArray}(x, y)

\S4method{query}{LabelArray}(x, y)

\S4method{query}{ShapeFrame}(x, y)

\S4method{query}{PointFrame}(x, y)
}
\arguments{
\item{x}{\code{SpatialData} element.}

\item{...}{optional arguments passed to and from other methods.}

\item{i}{for \code{SpatialData}, index or name of table to query.}

\item{y}{query specification; 
bounding box: length-4 numeric list with names 'xmin/xmax/ymin/ymax' 
(order is irrelevant); polygon: numeric matrix with ≥ 3 rows and 2 columns.}
}
\value{
same as input
}
\description{
Spatial queries serve to subset \code{SpatialData} elements 
according to a rectangular bounding box or arbitrary polygonal shapes. 
Queries rely on lesser-/greater-equal and \code{sf::st_intersects} for 
spatial operations (i.e., instances that intersect the query region 
in any way are kept). For circle shapes, radii are currently ignored
(i.e., a circle is kept if its centroid intersects the query region).
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")
sd <- readSpatialData(zs, tables=FALSE)

# helper for visualizing point coordinates
.xy <- \(.) data.frame(data(.)[c("x", "y")])

# bounding box
y <- list(xmin=11, xmax=44, ymin=22, ymax=55)
q <- query(p <- point(sd), y)

plot(.xy(p), asp=1)
points(.xy(q), col="red")
rect(y$xmin, y$ymin, y$xmax, y$ymax, border="blue")

# polygon
y <- rbind(c(20,10), c(50,30), c(20,50), c(30,30))
q <- query(p <- point(sd), y)

plot(.xy(p), asp=1)
points(.xy(q), col="red")
lines(rbind(y, y[1, ]), col="blue")

# shapes that intersect the query region are kept
y <- rbind(c(30,45), c(40,45), c(35,50))
t <- query(s <- shape(sd, 3), y)

require(sf, quietly=TRUE)
df <- st_coordinates(st_as_sf(data(s)))
fd <- st_coordinates(st_as_sf(data(t)))
plot(
  asp=1, xlim=c(15, 60), ylim=c(15, 60),
  rbind(y, y[1, ]), type="l", col="blue") 
foo <- by(df, df[, "L2"], \(x) points(x, type="b", col="black"))
foo <- by(fd, fd[, "L2"], \(x) points(x, type="b", col="red"))
}
