% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query across spatial layers}
\usage{
\S4method{query}{ZarrArray,LabelArray}(x, y, keep = NULL, crop = FALSE, ...)
}
\arguments{
\item{x, y}{\code{ImageArray}, \code{LabelArray}}

\item{keep}{Integer vector specifying the \code{y} values to retain.}

\item{crop}{Logical scalar specifying whether to crop empty space.}
}
\value{
\code{ImageArray}
}
\description{
...
}
\examples{
dir <- file.path("extdata", "blobs")
dir <- system.file(dir, package="SpatialData")
spd <- readSpatialData(dir)

img <- image(spd)
lab <- label(spd)

image(spd, "one") <- # keep any & don't crop
  query(img, lab, keep=NULL)
image(spd, "two") <- # keep one & do crop
  query(img, lab, keep=3, crop=TRUE)

plotImage(spd, i="one")
pile(
  plotLabel(spd), 
  plotImage(spd, i="two"))

}
\author{
Helena L. Crowell
}
