% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query,SpatialData-method}
\alias{query,ImageArray-method}
\alias{query,LabelArray-method}
\alias{query,ShapeFrame-method}
\alias{query,PointFrame-method}
\title{spatial queries}
\usage{
\S4method{query}{SpatialData}(x, j = NULL, ...)

\S4method{query}{ImageArray}(x, j, ...)

\S4method{query}{LabelArray}(x, ...)

\S4method{query}{ShapeFrame}(x, ...)

\S4method{query}{PointFrame}(x, j, ...)
}
\arguments{
\item{x}{\code{SpatialData} element.}

\item{j}{scalar character or integer; index or name of coordinate space.}

\item{...}{option arguments passed to and from other methods.}
}
\value{
same as input
}
\description{
spatial queries
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

image(x, "i") <- query(image(x), xmin=0, xmax=30, ymin=30, ymax=50)
plotSpatialData() + plotImage(x, "i")
}
