% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query,PointFrame-method}
\alias{query,ShapeFrame-method}
\title{spatial queries}
\usage{
\S4method{query}{PointFrame}(x, ...)

\S4method{query}{ShapeFrame}(x, ...)
}
\arguments{
\item{x}{\code{SpatialData} element}
}
\value{
...
}
\description{
spatial queries
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

p <- query(point(x), xmin=20, xmax=80, ymin=30, ymax=60)
dim(point(x, "p") <- p)
plotSpatialData() + plotPoint(x, "p")

s <- query(shape(x), xmin=20, xmax=80, ymin=30, ymax=60)
dim(shape(x, "s") <- s)
plotSpatialData() + plotShape(x, "s")
}
