% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImageArray.R
\name{readImageArray}
\alias{readImageArray}
\title{Read as `ImageArray`}
\usage{
readImageArray(path = ".", resolution = "0", ...)
}
\arguments{
\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{A \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}
}
\description{
...
}
\examples{
path <- "extdata/mibitof/images/point8_image/0"
path <- system.file(path, package = "SpatialData")
(ia <- readImageArray(path))

}
