% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPoints.R
\name{readPoints}
\alias{readPoints}
\title{Read `points` element}
\usage{
readPoints(path, ...)
}
\arguments{
\item{path}{A character string specifying
a .parquet file-containing directory.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
Arrow \code{\link{Dataset}}
}
\description{
...
}
\examples{
path <- "extdata/blobs/points/blobs_points"
path <- system.file(path, package = "SpatialData")
(ao <- readPoints(path))

}
\author{
Tim Treis
}
