% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readShapes}
\alias{readShapes}
\title{Read `shapes` element}
\usage{
readShapes(path, ...)
}
\arguments{
\item{path}{A character string specifying
the path to a \code{shapes/} subdirectory.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
\code{\link[S4Vectors]{DataFrame}}
}
\description{
...
}
\examples{
dir <- file.path("extdata", "blobs", "shapes", "blobs_polygons")
dir <- system.file(dir, package="SpatialData")
(sf <- readShapes(dir))
as.data.frame(sf)

}
\author{
Tim Treis, Helena L. Crowell
}
