% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readShapes.R
\name{readShapes}
\alias{readShapes}
\title{Read `shapes` element}
\usage{
readShapes(path, ...)
}
\arguments{
\item{path}{A character string specifying
the path to a `shapes/` subdirectory.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
\code{\link{DataFrame}}
}
\description{
...
}
\examples{
path <- file.path("extdata", "raccoon", "shapes", "circles")
path <- system.file(path, package="SpatialData")
(df <- readShapes(path))

}
\author{
Tim Treis
}
