% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpatialData.R
\name{readSpatialData}
\alias{readSpatialData}
\title{Read `SpatialData` OME-Zarr}
\usage{
readSpatialData(path, ...)
}
\arguments{
\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{An \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}
}
\description{
...
}
\examples{
path <- file.path("extdata", "mibitof")
path <- system.file(path, package = "SpatialData")
(spd <- readSpatialData(path))
images(spd)

path <- file.path("extdata", "blobs")
path <- system.file(path, package = "SpatialData")
(spd <- readSpatialData(path))
shapes(spd)

}
