% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readSpatialData}
\alias{readSpatialData}
\alias{readImage}
\alias{readLabel}
\alias{readPoint}
\alias{readShape}
\alias{readTable}
\title{Reading `SpatialData`}
\usage{
readImage(x)

readLabel(x)

readPoint(x)

readShape(x)

readTable(x)

readSpatialData(
  x,
  images = TRUE,
  labels = TRUE,
  points = TRUE,
  shapes = TRUE,
  tables = TRUE
)
}
\arguments{
\item{x}{For \code{readSpatialData},
path to a \code{SpatialData}.zarr store.
For \code{readImage/Label/Point/Shape/Table}, 
path to a \code{SpatialData} element.}

\item{images, labels, points, shapes, tables}{control which layers and elements to read;
FALSE to skip a layer, integer or character 
vector to read some element(s).}
}
\value{
\code{\link{SpatialData}} object, or one of its elements.
}
\description{
Reading `SpatialData`
}
\examples{
# package demo data
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
(sd <- readSpatialData(pa))

# view data available on OSN, 
# then retrieve & ingest one
Sys.setenv(AWS_REGION="us-east-1")
if (requireNamespace("paws")) sd_osn_list()
(sd <- sd_osn_load("merfish"))
}
