% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpatialData.R
\name{readSpatialData}
\alias{readSpatialData}
\title{Read `SpatialData` OME-Zarr}
\usage{
readSpatialData(path, ...)
}
\arguments{
\item{path}{A character string specifying the path to an
OME-Zarr file adhering to \code{SpatialData} specification.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\description{
...
}
\examples{
path <- file.path("extdata", "blobs")
path <- system.file(path, package="SpatialData")
(spd <- readSpatialData(path))

}
\author{
Constantin Ahlmann-Eltze, Helena L. Crowell
}
