% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readSpatialData}
\alias{readSpatialData}
\alias{readImage}
\alias{readLabel}
\alias{readPoint}
\alias{readShape}
\alias{readTable}
\title{Reading `SpatialData`}
\usage{
readImage(x)

readLabel(x)

readPoint(x)

readShape(x)

readTable(x)

readSpatialData(
  x,
  images = TRUE,
  labels = TRUE,
  points = TRUE,
  shapes = TRUE,
  tables = TRUE
)
}
\arguments{
\item{x}{For \code{readImage/Label/Point/Shape/Table}, 
path to a \code{SpatialData} element.
For \code{readSpatialData},
path to a \code{SpatialData}-.zarr store.}

\item{images, labels, points, shapes, tables}{Control which elements should be read for each layer.
The default, NULL, reads all elements; alternatively, may be FALSE 
to skip a layer, or a integer vector specifying which elements to read.}
}
\value{
\itemize{
\item{For element readers, a \code{ImageArray}, \code{LabelArray}, 
\code{PointFrame}, \code{ShapeFrame}, or \code{SingleCellExperiment}.}
\item{For \code{readSpatialData}, a \code{SpatialData}.}}
}
\description{
Reading `SpatialData`
}
\examples{
tf <- tempfile()
dir.create(tf)
base <- unzip_merfish_demo(tf)
(x <- readSpatialData(base))
}
