% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{SD-miscellaneous}
\alias{SD-miscellaneous}
\alias{show,SpatialData-method}
\alias{show,ImageArray-method}
\title{Miscellaneous `SpatialData` methods}
\usage{
\S4method{show}{SpatialData}(object)

\S4method{show}{ImageArray}(object)
}
\arguments{
\item{object}{\code{\link{SpatialData}} or \code{\link{ImageArray}} object.}
}
\value{
\code{NULL}
}
\description{
Miscellaneous methods for the \code{\link{SpatialData}}
and \code{\link{ImageArray}} classes that do not fit
into any other documentation category such as,
for example, show methods.
}
\examples{
path <- system.file("extdata", "raccoon", package="SpatialData")
(ia <- readImage(file.path(path, "images", "raccoon")))
(sd <- readSpatialData(path))
}
\author{
Helena L. Crowell
}
