% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sd_data}
\alias{sd_data}
\alias{sd_osn_list}
\alias{sd_osn_load}
\alias{all}
\alias{logic}
\alias{for}
\alias{finding,}
\alias{caching,}
\alias{loading}
\alias{an}
\alias{OSN-based}
\alias{dataset,}
\alias{hidden}
\title{Demo 'SpatialData'sets}
\usage{
sd_osn_list()

sd_osn_load(patt, target = tempfile(), cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{patt}{string sufficient to identify an OSN resource}

\item{target}{sting, defaults to \code{tempfile()}; use a different 
value if you wish to retain the unzipped .zarr store persistently.}

\item{cache}{like \code{BiocFileCache}}
}
\value{
\code{sd_osn_list}: character vector of .zip archives,
\code{sd_osn_load}: \code{\link{SpatialData}} object.
}
\description{
Demo 'SpatialData'sets
}
\note{
This function checks for stale element in cache and uses 
  \code{bfcupdate()} to rectify before retrieving from cache.
}
\examples{
# use 'paws::s3' to interrogate an NSF Open Storage Network 
# bucket for zipped .zarr archives for various platforms
Sys.setenv(AWS_REGION="us-east-1")
if (requireNamespace("paws")) sd_osn_list()

\donttest{
# retrieve dataset & ingest
(sd <- sd_osn_load("merfish"))
}
}
