% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sd_plot}
\alias{sd_plot}
\alias{sd_plot_image}
\title{Plotting `SpatialData`}
\usage{
sd_plot()
}
\value{
\code{ggplot}
}
\description{
Plotting `SpatialData`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

sd_plot() + 
  sd_plot_image(sd) +
  sd_plot_point(sd, c="x") 

pal <- c("cyan", "magenta", "gold")
sd_plot() + sd_plot_image(sd, c=pal)

}
\seealso{
\code{\link{sd_plot_image}}
\code{\link{sd_plot_point}}
\code{\link{sd_plot_shape}}
}
