% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-image.R
\name{sd_plot_image}
\alias{sd_plot_image}
\title{Plot `ImageArray`}
\usage{
sd_plot_image(
  x,
  i = 1,
  k = NULL,
  c = NULL,
  ch = NULL,
  cl = NULL,
  w = 800,
  h = 800
)
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i}{scalar integer or string; 
specifies which \code{images} to plot.}

\item{k}{resolution; if NULL (default), picking 
best for given \code{w}idth and \code{h}eight.}

\item{c}{character vector of colors to use;
if NULL (default), using \code{rainbow()}.}

\item{ch}{integer or character vector
specifying which channels to render.}

\item{cl}{list of channel-wise contrast limits.}

\item{w, h}{render width and height in pixel.}
}
\value{
\code{ggplot}
}
\description{
Plot `ImageArray`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

sd_plot() + sd_plot_image(sd)

# custom colors
pal <- c("cyan", "magenta", "gold")
sd_plot() + sd_plot_image(sd, c=pal)

}
