% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-label.R
\name{sd_plot_label}
\alias{sd_plot_label}
\title{Plot `LabelArray`}
\usage{
sd_plot_label(x, i = 1, c = NULL, k = NULL, w = 800, h = 800)
}
\arguments{
\item{x}{\code{\link{SpatialData}} object.}

\item{i}{scalar integer or string; 
specifies which \code{labels} to plot.}

\item{c}{character vector of colors to use;
if NULL (default), using \code{rainbow()}.}

\item{k}{resolution; if NULL (default), picking 
best for given \code{w}idth and \code{h}eight.}

\item{w, h}{render width and height in pixel.}
}
\value{
\code{ggplot}
}
\description{
Plot `LabelArray`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

sd_plot() + sd_plot_label(sd)
sd_plot() + sd_plot_label(sd, c="pink")
sd_plot() + sd_plot_label(sd, c=c("lavender", "blue"))

}
