% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-label.R
\name{sd_plot_label}
\alias{sd_plot_label}
\title{Plot `LabelArray`}
\usage{
sd_plot_label(x, i = 1, k = NULL, c = NULL, w = 800, h = 800)
}
\description{
Plot `LabelArray`
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)

devtools::load_all()
sd_plot() + sd_plot_label(sd)

}
