% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDattrs.R
\name{SDattrs}
\alias{SDattrs}
\alias{feature_key,list-method}
\alias{feature_key,PointFrame-method}
\alias{region_key,SingleCellExperiment-method}
\alias{region,SingleCellExperiment-method}
\alias{regions,SingleCellExperiment-method}
\alias{regions<-,SingleCellExperiment,character-method}
\alias{regions<-,SingleCellExperiment,NULL-method}
\alias{instance_key,list-method}
\alias{instance_key,sdFrame-method}
\alias{instance_key,LabelArray-method}
\alias{instance_key,SingleCellExperiment-method}
\alias{instance_key<-,SingleCellExperiment,character-method}
\alias{instances,LabelArray-method}
\alias{instances,PointFrame-method}
\alias{instances,ShapeFrame-method}
\alias{instances,SingleCellExperiment-method}
\alias{instances<-,SingleCellExperiment-method}
\title{\code{SpatialData} attributes}
\usage{
\S4method{feature_key}{list}(x)

\S4method{feature_key}{PointFrame}(x)

\S4method{region_key}{SingleCellExperiment}(x)

\S4method{region}{SingleCellExperiment}(x)

\S4method{regions}{SingleCellExperiment}(x)

\S4method{regions}{SingleCellExperiment,character}(x) <- value

\S4method{regions}{SingleCellExperiment,NULL}(x) <- value

\S4method{instance_key}{list}(x)

\S4method{instance_key}{sdFrame}(x)

\S4method{instance_key}{LabelArray}(x)

\S4method{instance_key}{SingleCellExperiment}(x)

\S4method{instance_key}{SingleCellExperiment,character}(x) <- value

\S4method{instances}{LabelArray}(x)

\S4method{instances}{PointFrame}(x)

\S4method{instances}{ShapeFrame}(x)

\S4method{instances}{SingleCellExperiment}(x)

\S4method{instances}{SingleCellExperiment}(x) <- value
}
\arguments{
\item{x}{depends on which attributes are available; 
specifically, \code{PointFrame} (\code{feature/instance_key}), or
\code{SingleCellExperiment} (\code{region}, \code{region/instance_key}),}
}
\value{
character string
}
\description{
\code{SpatialData} attributes
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x)

region(table(x))
region_key(table(x))

instance_key(point(x))
fk <- feature_key(point(x))
base::table(point(x)[[fk]])
}
