% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFrame.R
\name{ShapeFrame}
\alias{ShapeFrame}
\alias{data,ShapeFrame-method}
\alias{dim,ShapeFrame-method}
\alias{length,ShapeFrame-method}
\alias{names,ShapeFrame-method}
\alias{$,ShapeFrame-method}
\title{The `ShapeFrame` class}
\usage{
ShapeFrame(data = data.frame(), meta = Zattrs(), metadata = list(), ...)

\S4method{data}{ShapeFrame}(x)

\S4method{dim}{ShapeFrame}(x)

\S4method{length}{ShapeFrame}(x)

\S4method{names}{ShapeFrame}(x)

\S4method{$}{ShapeFrame}(x, name)
}
\arguments{
\item{data}{...}

\item{meta}{...}

\item{metadata}{....}
}
\value{
\code{ShapeFrame}
}
\description{
The `ShapeFrame` class
}
\examples{
x <- file.path("extdata", "merfish.zarr")

y <- file.path(x, "shapes", "cells")
(s <- readShape(system.file(y, package="SpatialData")))
plot(sf::st_as_sf(data(s)), cex=0.2)

y <- file.path(x, "shapes", "anatomical")
(s <- readShape(system.file(y, package="SpatialData")))
plot(sf::st_as_sf(data(s)), cex=0.2)

}
