% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFrame.R
\name{ShapeFrame}
\alias{ShapeFrame}
\alias{data,ShapeFrame-method}
\alias{dim,ShapeFrame-method}
\alias{length,ShapeFrame-method}
\alias{names,ShapeFrame-method}
\alias{$,ShapeFrame-method}
\title{The `ShapeFrame` class}
\usage{
ShapeFrame(data = data.frame(), meta = Zattrs(), metadata = list(), ...)

\S4method{data}{ShapeFrame}(x)

\S4method{dim}{ShapeFrame}(x)

\S4method{length}{ShapeFrame}(x)

\S4method{names}{ShapeFrame}(x)

\S4method{$}{ShapeFrame}(x, name)
}
\arguments{
\item{data}{\code{arrow}-derived table for on-disk,
\code{data.frame} for in-memory representation.}

\item{meta}{...}

\item{metadata}{....}
}
\value{
\code{ShapeFrame}
}
\description{
The `ShapeFrame` class
}
\examples{
tf <- tempfile()
dir.create(tf)
base <- unzip_merfish_demo(tf)
y <- file.path(base, "shapes", "cells")
(s <- readShape(y))
plot(sf::st_as_sf(data(s)), cex=0.2)

y <- file.path(base, "shapes", "anatomical")
(s <- readShape(y))
plot(sf::st_as_sf(data(s)), cex=0.2)

}
