% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFrame.R, R/transformations.R
\name{ShapeFrame}
\alias{ShapeFrame}
\alias{ShapeFrame-class}
\alias{coords,ShapeFrame-method}
\title{The `ShapeFrame` class}
\usage{
ShapeFrame(data = DataFrame(), metadata = list())

\S4method{coords}{ShapeFrame}(x)
}
\arguments{
\item{data}{A \code{\link{DataFrame}} of appropriate format.}

\item{metadata}{A list of metadata corresponding to .zattrs.}

\item{x}{An object of class \code{ShapeFrame}.}
}
\value{
\code{ShapeFrame}
}
\description{
A \code{\link{DataFrame}} with fixed structure
and .zattrs stored as metadata.
Each row corresponds to a shape, and is defined by its
\itemize{
\item{index: unique identifier}
\item{data:
  list of arrays containing xy-coordinates,
  and (if shapes are circles) radii.}
\item{type:
  character string specifying the geometry
  (\code{"circle"} or \code{"polygon"})}
}
}
\examples{
ShapeFrame()

}
\author{
Helena L. Crowell
}
