% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFrame.R
\name{ShapeFrame}
\alias{ShapeFrame}
\alias{dim,ShapeFrame-method}
\alias{length,ShapeFrame-method}
\alias{names,ShapeFrame-method}
\alias{$,ShapeFrame-method}
\alias{[,ShapeFrame,missing,ANY,ANY-method}
\alias{[,ShapeFrame,ANY,missing,ANY-method}
\alias{[,ShapeFrame,missing,missing,ANY-method}
\alias{[,ShapeFrame,numeric,numeric,ANY-method}
\title{The `ShapeFrame` class}
\usage{
ShapeFrame(data = data.frame(), meta = Zattrs(), metadata = list(), ...)

\S4method{dim}{ShapeFrame}(x)

\S4method{length}{ShapeFrame}(x)

\S4method{names}{ShapeFrame}(x)

\S4method{$}{ShapeFrame}(x, name)

\S4method{[}{ShapeFrame,missing,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ShapeFrame,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ShapeFrame,missing,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ShapeFrame,numeric,numeric,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{data}{\code{arrow}-derived table for on-disk,
\code{data.frame} for in-memory representation.}

\item{meta}{\code{\link{Zattrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{optional arguments passed to and from other methods.}

\item{x}{\code{ShapeFrame}}

\item{name}{character string for extraction (see \code{?base::`$`}).}
}
\value{
\code{ShapeFrame}
}
\description{
The `ShapeFrame` class
}
\examples{
tf <- tempfile()
dir.create(tf)
base <- unzip_merfish_demo(tf)
y <- file.path(base, "shapes", "cells")
(s <- readShape(y))
plot(sf::st_as_sf(data(s)), cex=0.2)

y <- file.path(base, "shapes", "anatomical")
(s <- readShape(y))
plot(sf::st_as_sf(data(s)), cex=0.2)

}
