% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialData-methods.R, R/SpatialData.R
\name{$,SpatialData-method}
\alias{$,SpatialData-method}
\alias{[[,SpatialData,ANY,ANY-method}
\alias{[[,SpatialData-method}
\alias{images,SpatialData-method}
\alias{image,SpatialData-method}
\alias{imageNames,SpatialData-method}
\alias{labels,SpatialData-method}
\alias{label,SpatialData-method}
\alias{labelNames,SpatialData-method}
\alias{shapes,SpatialData-method}
\alias{shape,SpatialData-method}
\alias{shapeNames,SpatialData-method}
\alias{points,SpatialData-method}
\alias{point,SpatialData-method}
\alias{pointNames,SpatialData-method}
\alias{table,SpatialData-method}
\alias{table<-,SpatialData,SingleCellExperiment_OR_NULL-method}
\alias{table<-,SpatialData,ANY-method}
\alias{elementNames,SpatialData-method}
\alias{element,SpatialData-method}
\alias{SpatialData}
\alias{SpatialData-class}
\alias{image}
\alias{images}
\alias{imageNames}
\alias{label}
\alias{labels}
\alias{labelNames}
\alias{shape}
\alias{shapes}
\alias{shapeNames}
\alias{point}
\alias{points}
\alias{pointNames}
\alias{element}
\alias{elementNames}
\alias{table}
\alias{table<-}
\title{The `SpatialData` class}
\usage{
\S4method{$}{SpatialData}(x, name)

\S4method{[[}{SpatialData,ANY,ANY}(x, i, j, ...)

\S4method{images}{SpatialData}(x)

\S4method{image}{SpatialData}(x, i = 1)

\S4method{imageNames}{SpatialData}(x)

\S4method{labels}{SpatialData}(x)

\S4method{label}{SpatialData}(x, i = 1)

\S4method{labelNames}{SpatialData}(x)

\S4method{shapes}{SpatialData}(x)

\S4method{shape}{SpatialData}(x, i = 1)

\S4method{shapeNames}{SpatialData}(x)

\S4method{points}{SpatialData}(x)

\S4method{point}{SpatialData}(x, i = 1)

\S4method{pointNames}{SpatialData}(x)

\S4method{table}{SpatialData}(x)

\S4method{table}{SpatialData,SingleCellExperiment_OR_NULL}(x) <- value

\S4method{table}{SpatialData,ANY}(x) <- value

\S4method{elementNames}{SpatialData}(x)

\S4method{element}{SpatialData}(x, elementName = elementNames(x)[1], i = 1, ...)

SpatialData(images, labels, shapes, points, table)
}
\arguments{
\item{x}{A \code{SpatialData} object.}

\item{i}{Entity of the respective element to extract;
can be an integer index or character string
(one of \code{eNames(x)}, where \code{e}
is the specified \code{elementName}).}

\item{j}{Ignored.}

\item{...}{Further arguments to be passed to or from other methods.}

\item{value}{Object of appropriate type; see respective elements.}

\item{elementName, name}{A character string
specifying the type of element to extract;
should be one of \code{"images"}, \code{"labels"},
\code{"shapes"}, \code{"points"}, or \code{"table"}.}

\item{images}{A list of \code{\link{ImageArray}}s.}

\item{labels}{A list of \code{\link{ImageArray}}s.}

\item{shapes}{A list of \code{\link{DataFrame}}s.}

\item{points}{A list of Arrow \code{\link{Dataset}}s.}

\item{table}{A \code{SingleCellExperiment}s.}
}
\value{
\itemize{
\item \code{images/labels/shapes/points}
  return a list of entities of the corresponding element.
\item \code{image/label/shape/point}
  return a single entity of the corresponding type.
\item \code{image/label/shape/pointNames}
  return a character string of available
  entities of the corresponding element.
}
}
\description{
...
}
\examples{
path <- file.path("extdata", "blobs")
path <- system.file(path, package="SpatialData")
(spd <- readSpatialData(path))

# accessors
imageNames(spd)
image(spd, "blobs_image")
spd$images$blobs_image

(sce <- table(spd))

}
\author{
Constantin Ahlmann-Eltze, Helena L. Crowell
}
