% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialData-methods.R, R/SpatialData.R,
%   R/Zattrs.R
\name{$,SpatialData-method}
\alias{$,SpatialData-method}
\alias{[[,SpatialData,ANY,ANY-method}
\alias{images,SpatialData-method}
\alias{labels,SpatialData-method}
\alias{shapes,SpatialData-method}
\alias{points,SpatialData-method}
\alias{tables,SpatialData-method}
\alias{elements,SpatialData-method}
\alias{image,SpatialData-method}
\alias{label,SpatialData-method}
\alias{shape,SpatialData-method}
\alias{point,SpatialData-method}
\alias{table,SpatialData-method}
\alias{element,SpatialData-method}
\alias{images<-,SpatialData,list-method}
\alias{labels<-,SpatialData,list-method}
\alias{shapes<-,SpatialData,list-method}
\alias{points<-,SpatialData,list-method}
\alias{tables<-,SpatialData,list-method}
\alias{image<-,SpatialData,character,ImageArray-method}
\alias{label<-,SpatialData,character,LabelArray-method}
\alias{shape<-,SpatialData,character,ShapeFrame-method}
\alias{point<-,SpatialData,character,PointFrame-method}
\alias{table<-,SpatialData,character,SingleCellExperiment-method}
\alias{image<-,SpatialData,missing,ImageArray-method}
\alias{label<-,SpatialData,missing,LabelArray-method}
\alias{shape<-,SpatialData,missing,ShapeFrame-method}
\alias{point<-,SpatialData,missing,PointFrame-method}
\alias{table<-,SpatialData,missing,SingleCellExperiment-method}
\alias{image<-,SpatialData,numeric,ImageArray-method}
\alias{label<-,SpatialData,numeric,LabelArray-method}
\alias{shape<-,SpatialData,numeric,ShapeFrame-method}
\alias{point<-,SpatialData,numeric,PointFrame-method}
\alias{table<-,SpatialData,numeric,SingleCellExperiment-method}
\alias{image<-,SpatialData,ANY,NULL-method}
\alias{label<-,SpatialData,ANY,NULL-method}
\alias{shape<-,SpatialData,ANY,NULL-method}
\alias{point<-,SpatialData,ANY,NULL-method}
\alias{table<-,SpatialData,ANY,NULL-method}
\alias{image<-,SpatialData,ANY,ANY-method}
\alias{label<-,SpatialData,ANY,ANY-method}
\alias{shape<-,SpatialData,ANY,ANY-method}
\alias{point<-,SpatialData,ANY,ANY-method}
\alias{table<-,SpatialData,ANY,ANY-method}
\alias{imageNames,SpatialData-method}
\alias{labelNames,SpatialData-method}
\alias{shapeNames,SpatialData-method}
\alias{pointNames,SpatialData-method}
\alias{tableNames,SpatialData-method}
\alias{elementNames,SpatialData-method}
\alias{imageNames<-,SpatialData,character-method}
\alias{labelNames<-,SpatialData,character-method}
\alias{shapeNames<-,SpatialData,character-method}
\alias{pointNames<-,SpatialData,character-method}
\alias{tableNames<-,SpatialData,character-method}
\alias{SpatialData}
\alias{SpatialData-class}
\alias{[[,SpatialData-method}
\alias{element}
\alias{elements}
\alias{elementNames}
\alias{image}
\alias{label}
\alias{shape}
\alias{point}
\alias{table}
\alias{images}
\alias{labels}
\alias{shapes}
\alias{points}
\alias{tables}
\alias{image<-}
\alias{label<-}
\alias{shape<-}
\alias{point<-}
\alias{table<-}
\alias{images<-}
\alias{labels<-}
\alias{shapes<-}
\alias{points<-}
\alias{tables<-}
\alias{imageNames}
\alias{labelNames}
\alias{shapeNames}
\alias{pointNames}
\alias{tableNames}
\alias{imageNames<-}
\alias{labelNames<-}
\alias{shapeNames<-}
\alias{pointNames<-}
\alias{tableNames<-}
\alias{image<-,SpatialData,1}
\alias{image<-,SpatialData,character}
\alias{images<-,SpatialData,list}
\alias{label<-,SpatialData,1}
\alias{label<-,SpatialData,character}
\alias{labels<-,SpatialData,list}
\alias{shape<-,SpatialData,1}
\alias{shape<-,SpatialData,character}
\alias{shapes<-,SpatialData,list}
\alias{point<-,SpatialData,1}
\alias{point<-,SpatialData,character}
\alias{points<-,SpatialData,list}
\alias{Zattrs}
\title{The `SpatialData` class}
\usage{
\S4method{$}{SpatialData}(x, name)

\S4method{[[}{SpatialData,ANY,ANY}(x, i, j, ...)

\S4method{images}{SpatialData}(x)

\S4method{labels}{SpatialData}(x)

\S4method{shapes}{SpatialData}(x)

\S4method{points}{SpatialData}(x)

\S4method{tables}{SpatialData}(x)

\S4method{elements}{SpatialData}(x)

\S4method{image}{SpatialData}(x, i = 1)

\S4method{label}{SpatialData}(x, i = 1)

\S4method{shape}{SpatialData}(x, i = 1)

\S4method{point}{SpatialData}(x, i = 1)

\S4method{table}{SpatialData}(x, i = 1)

\S4method{element}{SpatialData}(x, e = 1, i = 1)

\S4method{images}{SpatialData,list}(x) <- value

\S4method{labels}{SpatialData,list}(x) <- value

\S4method{shapes}{SpatialData,list}(x) <- value

\S4method{points}{SpatialData,list}(x) <- value

\S4method{tables}{SpatialData,list}(x) <- value

\S4method{image}{SpatialData,character,ImageArray}(x, i) <- value

\S4method{label}{SpatialData,character,LabelArray}(x, i) <- value

\S4method{shape}{SpatialData,character,ShapeFrame}(x, i) <- value

\S4method{point}{SpatialData,character,PointFrame}(x, i) <- value

\S4method{table}{SpatialData,character,SingleCellExperiment}(x, i) <- value

\S4method{image}{SpatialData,missing,ImageArray}(x, i) <- value

\S4method{label}{SpatialData,missing,LabelArray}(x, i) <- value

\S4method{shape}{SpatialData,missing,ShapeFrame}(x, i) <- value

\S4method{point}{SpatialData,missing,PointFrame}(x, i) <- value

\S4method{table}{SpatialData,missing,SingleCellExperiment}(x, i) <- value

\S4method{image}{SpatialData,numeric,ImageArray}(x, i = 1) <- value

\S4method{label}{SpatialData,numeric,LabelArray}(x, i = 1) <- value

\S4method{shape}{SpatialData,numeric,ShapeFrame}(x, i = 1) <- value

\S4method{point}{SpatialData,numeric,PointFrame}(x, i = 1) <- value

\S4method{table}{SpatialData,numeric,SingleCellExperiment}(x, i = 1) <- value

\S4method{image}{SpatialData,ANY,`NULL`}(x, i) <- value

\S4method{label}{SpatialData,ANY,`NULL`}(x, i) <- value

\S4method{shape}{SpatialData,ANY,`NULL`}(x, i) <- value

\S4method{point}{SpatialData,ANY,`NULL`}(x, i) <- value

\S4method{table}{SpatialData,ANY,`NULL`}(x, i) <- value

\S4method{image}{SpatialData,ANY,ANY}(x, i) <- value

\S4method{label}{SpatialData,ANY,ANY}(x, i) <- value

\S4method{shape}{SpatialData,ANY,ANY}(x, i) <- value

\S4method{point}{SpatialData,ANY,ANY}(x, i) <- value

\S4method{table}{SpatialData,ANY,ANY}(x, i) <- value

\S4method{imageNames}{SpatialData}(x)

\S4method{labelNames}{SpatialData}(x)

\S4method{shapeNames}{SpatialData}(x)

\S4method{pointNames}{SpatialData}(x)

\S4method{tableNames}{SpatialData}(x)

\S4method{elementNames}{SpatialData}(x)

\S4method{imageNames}{SpatialData,character}(x) <- value

\S4method{labelNames}{SpatialData,character}(x) <- value

\S4method{shapeNames}{SpatialData,character}(x) <- value

\S4method{pointNames}{SpatialData,character}(x) <- value

\S4method{tableNames}{SpatialData,character}(x) <- value

SpatialData(images, labels, shapes, points, tables)

Zattrs(x = list())
}
\arguments{
\item{x}{A \code{SpatialData} object.}

\item{i}{Entity of the respective element to extract;
can be an integer index or character string
(one of \code{eNames(x)}, where \code{e}
is the specified \code{elementName}).}

\item{j}{Ignored.}

\item{...}{Further arguments to be passed to or from other methods.}

\item{e, name}{A character string specifying the type of element to extract;
should be one of \code{"images","labels","shapes","points","tables"}.}

\item{value}{Object of appropriate type; see respective elements.}

\item{images}{A list of \code{\link{ImageArray}}s.}

\item{labels}{A list of \code{\link{ImageArray}}s.}

\item{shapes}{A list of \code{\link{DataFrame}}s.}

\item{points}{A list of Arrow \code{\link{Dataset}}s.}

\item{tables}{A list of \code{SingleCellExperiment}s.}
}
\value{
\itemize{
\item \code{images/labels/shapes/points}
  return a list of entities of the corresponding element.
\item \code{image/label/shape/point}
  return a single entity of the corresponding type.
\item \code{image/label/shape/pointNames}
  return a character string of available
  entities of the corresponding element.
}
}
\description{
...
}
\examples{
dir <- file.path("extdata", "blobs")
dir <- system.file(dir, package="SpatialData")
(spd <- readSpatialData(dir))

# accessors
imageNames(spd)
image(spd, "blobs_image")
spd$images$blobs_image

(sce <- table(spd))

}
\author{
Constantin Ahlmann-Eltze, Helena L. Crowell
}
