% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/SpatialData.R
\name{images,SpatialData-method}
\alias{images,SpatialData-method}
\alias{labels,SpatialData-method}
\alias{shapes,SpatialData-method}
\alias{points,SpatialData-method}
\alias{SpatialData}
\title{The `SpatialData` class}
\usage{
\S4method{images}{SpatialData}(x)

\S4method{labels}{SpatialData}(x)

\S4method{shapes}{SpatialData}(x)

\S4method{points}{SpatialData}(x)

SpatialData(images, labels, shapes, points, ...)
}
\arguments{
\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{A \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}
}
\description{
...
}
\examples{
path <- "extdata/mibitof/images/point8_image/mibitof"
path <- system.file(path, package = "SpatialData")
(spd <- readSpatialData(path))

}
