% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/SpatialData.R
\name{images,SpatialData-method}
\alias{images,SpatialData-method}
\alias{labels,SpatialData-method}
\alias{shapes,SpatialData-method}
\alias{points,SpatialData-method}
\alias{label,SpatialData-method}
\alias{shape,SpatialData-method}
\alias{point,SpatialData-method}
\alias{imageNames,SpatialData-method}
\alias{labelNames,SpatialData-method}
\alias{shapeNames,SpatialData-method}
\alias{pointNames,SpatialData-method}
\alias{elementNames,SpatialData-method}
\alias{element,SpatialData-method}
\alias{SpatialData}
\title{The `SpatialData` class}
\usage{
\S4method{images}{SpatialData}(x)

\S4method{labels}{SpatialData}(x)

\S4method{shapes}{SpatialData}(x)

\S4method{points}{SpatialData}(x)

\S4method{images}{SpatialData}(x)

\S4method{label}{SpatialData}(x, i = 1)

\S4method{shape}{SpatialData}(x, i = 1)

\S4method{point}{SpatialData}(x, i = 1)

\S4method{imageNames}{SpatialData}(x)

\S4method{labelNames}{SpatialData}(x)

\S4method{shapeNames}{SpatialData}(x)

\S4method{pointNames}{SpatialData}(x)

\S4method{elementNames}{SpatialData}(x)

\S4method{element}{SpatialData}(x, elementName = elementNames(x)[1], which = 1, ...)

SpatialData(images, labels, shapes, points, ...)
}
\arguments{
\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{A \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}
}
\description{
...
}
\examples{
path <- file.path("extdata", "mibitof")
path <- system.file(path, package = "SpatialData")
(spd <- readSpatialData(path))
images(spd)

}
