% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{SpatialData}
\alias{SpatialData}
\alias{images}
\alias{labels}
\alias{points}
\alias{shapes}
\alias{tables}
\alias{ImageArray}
\alias{LabelArray}
\alias{PointFrame}
\alias{ShapeFrame}
\alias{feature_key}
\alias{instance_key}
\alias{zattrs}
\title{The `SpatialData` class}
\usage{
SpatialData(
  images = list(),
  labels = list(),
  points = list(),
  shapes = list(),
  tables = list()
)

ImageArray(data = list(), zattrs = Zattrs(), metadata = list())

LabelArray(data = list(), zattrs = Zattrs(), metadata = list())

PointFrame(data = NULL, zattrs = Zattrs(), metadata = list())

feature_key(x, ...)

instance_key(x, ...)

ShapeFrame(data = NULL, zattrs = Zattrs(), metadata = list())

zattrs(x, ...)
}
\arguments{
\item{images}{list of \code{\link{ImageArray}}s}

\item{labels}{list of \code{\link{LabelArray}}s}

\item{points}{list of \code{\link{PointFrame}}s}

\item{shapes}{list of \code{\link{ShapeFrame}}s}

\item{tables}{list of \code{SingleCellExperiment}s}
}
\value{
\code{SpatialData}
}
\description{
The `SpatialData` class
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
(sd <- readSpatialData(pa))

names(sd)
sd@images[[1]]
sd["shapes", 2]
region(sd, "table")

}
