% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SpatialData.R, R/methods.R
\docType{class}
\name{SpatialData-class}
\alias{SpatialData-class}
\alias{.SpatialData}
\alias{SpatialData}
\alias{data}
\alias{meta}
\alias{image}
\alias{images}
\alias{image<-}
\alias{images<-}
\alias{imageNames}
\alias{label}
\alias{labels}
\alias{label<-}
\alias{labels<-}
\alias{labelNames}
\alias{point}
\alias{points}
\alias{point<-}
\alias{points<-}
\alias{pointNames}
\alias{shape}
\alias{shapes}
\alias{shape<-}
\alias{shapes<-}
\alias{shapeNames}
\alias{table}
\alias{tables}
\alias{table<-}
\alias{tables<-}
\alias{tableNames}
\alias{$,SpatialData-method}
\alias{[[,SpatialData,numeric,ANY-method}
\alias{[[,SpatialData,character,ANY-method}
\alias{[,SpatialData,ANY,ANY,ANY-method}
\alias{data,SpatialDataElement-method}
\alias{meta,SpatialDataElement-method}
\alias{rownames,SpatialData-method}
\alias{colnames,SpatialData-method}
\alias{f}
\alias{[[<-,SpatialData,numeric,ANY,ANY-method}
\alias{[[<-,SpatialData,character,ANY,ANY-method}
\alias{g}
\title{The `SpatialData` class}
\usage{
SpatialData(images, labels, points, shapes, tables)

\S4method{$}{SpatialData}(x, name)

\S4method{[[}{SpatialData,numeric,ANY}(x, i, j, ...)

\S4method{[[}{SpatialData,character,ANY}(x, i, j, ...)

\S4method{[}{SpatialData,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{data}{SpatialDataElement}(x)

\S4method{meta}{SpatialDataElement}(x)

\S4method{rownames}{SpatialData}(x)

\S4method{colnames}{SpatialData}(x)

f(.)

f(.)

f(.)

\S4method{[[}{SpatialData,numeric,ANY,ANY}(x, i) <- value

\S4method{[[}{SpatialData,character,ANY,ANY}(x, i) <- value

f(.)

f(.)

f(.)

f(.)

f(.)

g(.)
}
\arguments{
\item{images}{list of \code{\link{ImageArray}}s}

\item{labels}{list of \code{\link{LabelArray}}s}

\item{points}{list of \code{\link{PointFrame}}s}

\item{shapes}{list of \code{\link{ShapeFrame}}s}

\item{tables}{list of \code{SingleCellExperiment}s}

\item{x}{\code{SpatialData}}

\item{name}{character string for extraction (see \code{?base::`$`}).}

\item{i, j}{character string, scalar or vector of indices
specifying the element to extract from a given layer.}

\item{...}{optional arguments passed to and from other methods.}

\item{drop}{ignored.}

\item{value}{(list of) element(s) with layer-compliant object(s), 
or NULL/\code{list()} to remove an element/layer completely.}
}
\value{
\code{SpatialData}
}
\description{
...
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
(x <- readSpatialData(x, anndataR=TRUE))

# subsetting
# layers are taken in order of appearance
# (images, labels, points, shapes, tables)
x[-4] # drop layer
x[4, -2] # drop element
x["shapes", c(1, 3)] # subset layer
x[c(1, 2), list(1, c(1, 2))] # multiple

}
