% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/SpatialData.R,
%   R/aggregateImage.R
\name{images,SpatialData-method}
\alias{images,SpatialData-method}
\alias{labels,SpatialData-method}
\alias{shapes,SpatialData-method}
\alias{points,SpatialData-method}
\alias{image,SpatialData-method}
\alias{label,SpatialData-method}
\alias{shape,SpatialData-method}
\alias{point,SpatialData-method}
\alias{imageNames,SpatialData-method}
\alias{labelNames,SpatialData-method}
\alias{shapeNames,SpatialData-method}
\alias{pointNames,SpatialData-method}
\alias{elementNames,SpatialData-method}
\alias{element,SpatialData-method}
\alias{$,SpatialData-method}
\alias{SpatialData}
\alias{aggregateImage}
\title{The `SpatialData` class}
\usage{
\S4method{images}{SpatialData}(x)

\S4method{labels}{SpatialData}(x)

\S4method{shapes}{SpatialData}(x)

\S4method{points}{SpatialData}(x)

\S4method{image}{SpatialData}(x, i = 1)

\S4method{label}{SpatialData}(x, i = 1)

\S4method{shape}{SpatialData}(x, i = 1)

\S4method{point}{SpatialData}(x, i = 1)

\S4method{imageNames}{SpatialData}(x)

\S4method{labelNames}{SpatialData}(x)

\S4method{shapeNames}{SpatialData}(x)

\S4method{pointNames}{SpatialData}(x)

\S4method{elementNames}{SpatialData}(x)

\S4method{element}{SpatialData}(x, elementName = elementNames(x)[1], which = 1, ...)

\S4method{$}{SpatialData}(x, name)

SpatialData(images, labels, shapes, points, ...)

aggregateImage(x, which = 1)
}
\arguments{
\item{x}{A \code{SpatialData} object.}

\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{A \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}
}
\value{
\itemize{
\item \code{images/labels/shapes/points}
  return a list of entities of the corresponding element.
\item \code{image/label/shape/point}
  return a single entitiy of the corresponing type.
\item \code{image/label/shape/pointNames}
  return a character string of available 
  entities of the corresponding element.
}
}
\description{
...

...
}
\examples{
path <- file.path("extdata", "mibitof")
path <- system.file(path, package = "SpatialData")
(spd <- readSpatialData(path))
images(spd)

library(ggplot2)
library(SingleCellExperiment)
path <- file.path("extdata", "mibitof")
path <- system.file(path, package = "SpatialData")
spd <- readSpatialData(path)
sce <- aggregateImage(spd)
cd <- data.frame(colData(sce), z = assay(sce)[1, ])
ggplot(cd, aes(x, y, col = z)) + geom_point() +
  scale_color_viridis_c() + scale_y_reverse()

}
