% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialData.R
\name{SpatialData}
\alias{SpatialData}
\title{The `SpatialData` class}
\usage{
SpatialData(images, labels, shapes, points, ...)
}
\arguments{
\item{...}{Further arguments to be passed to or from other methods.}

\item{data}{A \code{array} or \code{\link[S4Arrays]{Array}}.}

\item{metadata}{A \code{list}.}
}
\description{
...
}
\examples{
dir <- "extdata/mibitof/images/point8_image"
zarr <- system.file(file.path(dir, "0"), package = "SpatialData")
json <- system.file(file.path(dir, ".zattrs"), package = "SpatialData")

library(Rarr)
library(jsonlite)

za <- read_zarr_array(zarr)
md <- fromJSON(json)
(ia <- ImageArray(za, md))

}
