% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{SpatialData}
\alias{SpatialData}
\alias{images}
\alias{labels}
\alias{points}
\alias{shapes}
\alias{tables}
\alias{ImageArray}
\alias{LabelArray}
\alias{PointFrame}
\alias{ShapeFrame}
\alias{region}
\alias{feature_key}
\alias{instance_key}
\alias{zattrs}
\title{The `SpatialData` class}
\usage{
SpatialData(
  images = list(),
  labels = list(),
  points = list(),
  shapes = list(),
  tables = list()
)

region(x, ...)

ImageArray(data = list(), zattrs = Zattrs(), metadata = list())

LabelArray(data = list(), zattrs = Zattrs(), metadata = list())

PointFrame(data = NULL, zattrs = Zattrs(), metadata = list())

feature_key(x, ...)

instance_key(x, ...)

ShapeFrame(data = NULL, zattrs = Zattrs(), metadata = list())

zattrs(x, ...)
}
\arguments{
\item{images}{list of \code{\link{ImageArray}}s}

\item{labels}{list of \code{\link{LabelArray}}s}

\item{points}{list of \code{\link{PointFrame}}s}

\item{shapes}{list of \code{\link{ShapeFrame}}s}

\item{tables}{list of \code{SingleCellExperiment}s}

\item{x}{\code{SpatialData} object.}

\item{data}{For \code{Image/LabelArray}s, a list of \code{ZarrArray}s;
For \code{Shape/PointFrame}s, an \code{arrow} Table or derivative.}

\item{zattrs}{\code{Zattrs} objects; used to represent .zattrs.}

\item{metadata}{(option) list of free-form extra data.}
}
\value{
\code{SpatialData}
}
\description{
The `SpatialData` class
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
(sd <- readSpatialData(pa))

names(sd)
sd@images[[1]]
sd[-4, ][, 1]
sd[1:2, c(1,1)]
sd["shapes", c(1,3)]
region(sd, "table")

}
