% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdAttrs.R
\name{SpatialDataAttrs}
\alias{SpatialDataAttrs}
\alias{region}
\alias{region<-}
\alias{regions}
\alias{regions<-}
\alias{instances}
\alias{instances<-}
\alias{region_key}
\alias{region_key<-}
\alias{feature_key}
\alias{feature_key<-}
\alias{instance_key}
\alias{instance_key<-}
\alias{$,SpatialDataAttrs-method}
\alias{feature_key,SpatialDataPoint-method}
\alias{feature_key,SpatialDataAttrs-method}
\alias{feature_key<-,SpatialDataAttrs,character-method}
\alias{region_key,SingleCellExperiment-method}
\alias{region,SingleCellExperiment-method}
\alias{regions,SingleCellExperiment-method}
\alias{regions<-,SingleCellExperiment,character-method}
\alias{regions<-,SingleCellExperiment,NULL-method}
\alias{instance_key,list-method}
\alias{instance_key,SingleCellExperiment-method}
\alias{instance_key,SpatialDataFrame-method}
\alias{instance_key,SpatialDataLabel-method}
\alias{instance_key<-,SpatialDataAttrs,character-method}
\alias{instance_key<-,SingleCellExperiment,character-method}
\alias{instances,SpatialDataLabel-method}
\alias{instances,SpatialDataPoint-method}
\alias{instances,SpatialDataShape-method}
\alias{instances,SingleCellExperiment-method}
\alias{instances<-,SingleCellExperiment-method}
\title{The `SpatialDataAttrs` class}
\usage{
SpatialDataAttrs(
  x,
  type = c("image", "label", "point", "shape"),
  trans = NULL,
  ver = NULL,
  dim = 2,
  nch = 3,
  ...
)

\S4method{$}{SpatialDataAttrs}(x, name)

\S4method{feature_key}{SpatialDataPoint}(x)

\S4method{feature_key}{SpatialDataAttrs}(x)

\S4method{feature_key}{SpatialDataAttrs,character}(x) <- value

\S4method{region_key}{SingleCellExperiment}(x)

\S4method{region}{SingleCellExperiment}(x)

\S4method{regions}{SingleCellExperiment}(x)

\S4method{regions}{SingleCellExperiment,character}(x) <- value

\S4method{regions}{SingleCellExperiment,NULL}(x) <- value

\S4method{instance_key}{list}(x)

\S4method{instance_key}{SingleCellExperiment}(x)

\S4method{instance_key}{SpatialDataFrame}(x)

\S4method{instance_key}{SpatialDataLabel}(x)

\S4method{instance_key}{SpatialDataAttrs,character}(x) <- value

\S4method{instance_key}{SingleCellExperiment,character}(x) <- value

\S4method{instances}{SpatialDataLabel}(x)

\S4method{instances}{SpatialDataPoint}(x)

\S4method{instances}{SpatialDataShape}(x)

\S4method{instances}{SingleCellExperiment}(x)

\S4method{instances}{SingleCellExperiment}(x) <- value
}
\arguments{
\item{x}{element or list extracted from a OME-NGFF compliant .zattrs file.}

\item{type}{character string; either "image", "label", "point" or "shape"}

\item{trans}{list of coordinate transformations; defaults to identity only.}

\item{ver}{character string; specifies the SpatialData version to comply with.}

\item{dim}{scalar integer in 2-4;
number of dimensions: 2 = XY, 3 adds Z, 4 adds T (time); 
when \code{type="image"}, C (channel) will be added (for any \code{dim}).}

\item{nch}{scalar integer; how many channels should there be?
(ignored unless \code{type="shape"} or \code{type="point"}, and 
\code{label=FALSE}).}

\item{...}{additional attributes (e.g., version, feature_key).}

\item{name}{character string for extraction (see ?base::`$`).}

\item{value}{character string (for one \code{region} and \code{_key}s), 
or vector (for many \code{region}s, \code{instances} and \code{regions}).}
}
\value{
character string
}
\description{
The `SpatialDataAttrs` class
}
\details{
When \code{x} is a spatial element, the following applies:
\code{SpatialDataFrame}: \code{feature/instance_key},
\code{SingleCellExperiment}: \code{region}, \code{region/instance_key}.

When missing \code{x}, \code{SpatialDataAttrs} will generate a valid object 
with default axes (image/label: cyx, point/shape: xy) and transformations 
(identify) according to the specified type.
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="spatialdataR")
x <- readSpatialData(x)

# tables
region(table(x))
region_key(table(x))

# points
instance_key(point(x))
fk <- feature_key(point(x))
base::table(point(x)[[fk]])

# transformations
(z <- meta(label(x)))
CTname(z)
CTtype(z)
CTdata(z, "scale")

# constructor
SpatialDataAttrs(type="point")
SpatialDataAttrs(type="shape")
SpatialDataAttrs(type="image", nch=7)
SpatialDataAttrs(type="label", dim=3)

}
