% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{spdzPath}
\alias{spdzPath}
\title{obtain path to cached zip archive of SpatialData zarr}
\usage{
spdzPath(cache = BiocFileCache::BiocFileCache(), zipname, source)
}
\arguments{
\item{cache}{inherits from \code{BiocFileCache::BiocFileCache()}}

\item{zipname}{character(1) name of zip archive to find}

\item{source}{character(1) one of "biocOSN", "sandbox", "local"}
}
\description{
obtain path to cached zip archive of SpatialData zarr
}
\examples{
spdzPath(zipname="merfish.zarr.zip", source="biocOSN")
}
