% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\alias{scale}
\alias{rotate}
\alias{translation}
\alias{flip}
\alias{flop}
\alias{mirror}
\alias{mirror,sdArray-method}
\alias{flip,sdArray-method}
\alias{flop,sdArray-method}
\alias{scale,sdArray,numeric-method}
\alias{rotate,sdArray,numeric-method}
\alias{scale,PointFrame,numeric-method}
\alias{rotate,PointFrame,numeric-method}
\alias{translation,PointFrame,numeric-method}
\alias{scale,ShapeFrame,numeric-method}
\alias{rotate,ShapeFrame,numeric-method}
\alias{translation,ShapeFrame,numeric-method}
\title{Transformations}
\usage{
\S4method{mirror}{sdArray}(x, t = c("v", "h"), k = 1, ...)

\S4method{flip}{sdArray}(x, k = 1, ...)

\S4method{flop}{sdArray}(x, k = 1, ...)

\S4method{scale}{sdArray,numeric}(x, t, k = 1, ...)

\S4method{rotate}{sdArray,numeric}(x, t, k = 1, ...)

\S4method{scale}{PointFrame,numeric}(x, t, ...)

\S4method{rotate}{PointFrame,numeric}(x, t, ...)

\S4method{translation}{PointFrame,numeric}(x, t, ...)

\S4method{scale}{ShapeFrame,numeric}(x, t, ...)

\S4method{rotate}{ShapeFrame,numeric}(x, t, ...)

\S4method{translation}{ShapeFrame,numeric}(x, t, ...)
}
\arguments{
\item{x}{\code{SpatialData} element}

\item{t}{transformation data; exceptions: for \code{mirror}, controls
whether to perform \bold{v}ertical or \bold{h}orizontal reflection;
no data is needed for \code{flip} (\bold{v}) and \code{flop} (\bold{h}).}

\item{k}{scalar index specifying which scale to use; 
\code{Inf} to use lowest available resolution;
only applies to \code{sdArray}s (images, labels).}

\item{...}{option arguments passed to and from other methods.}

\item{j}{scalar character or numeric; 
name or index of coordinate space.}
}
\description{
Transformations
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# image
y <- x
image(y) <- scale(image(y), c(1, 1, 1/3))
dim(image(x))
dim(image(y))
  
# point
y <- x
point(y, "rot") <- rotate(point(y), 20)
point(y, "wide") <- scale(point(y), c(1.2, 1))

xy0 <- as.data.frame(point(y))
xy1 <- as.data.frame(point(y, "rot"))
xy2 <- as.data.frame(point(y, "wide"))

plot(xy0[, c(1, 2)], asp=1)
points(xy1[, c(1, 2)], col=2)
points(xy2[, c(1, 2)], col=4)
  
# shape
y <- x
shape(y, "rot") <- rotate(shape(y), 5)
shape(y, "wide") <- scale(shape(y), c(1.2, 1))
shape(y, "left") <- translation(shape(y), c(-5, 0))
y["shapes", c("rot", "wide", "left")]
}
