% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\alias{scale}
\alias{rotate}
\alias{translation}
\alias{scale,ImageArray,numeric-method}
\alias{scale,PointFrame,numeric-method}
\alias{rotate,PointFrame,numeric-method}
\alias{translation,PointFrame,numeric-method}
\alias{scale,ShapeFrame,numeric-method}
\alias{rotate,ShapeFrame,numeric-method}
\alias{translation,ShapeFrame,numeric-method}
\title{Transformations}
\usage{
\S4method{scale}{ImageArray,numeric}(x, j, t, ...)

\S4method{scale}{PointFrame,numeric}(x, t, ...)

\S4method{rotate}{PointFrame,numeric}(x, t, ...)

\S4method{translation}{PointFrame,numeric}(x, t, ...)

\S4method{scale}{ShapeFrame,numeric}(x, t, ...)

\S4method{rotate}{ShapeFrame,numeric}(x, t, ...)

\S4method{translation}{ShapeFrame,numeric}(x, t, ...)
}
\arguments{
\item{x}{\code{SpatialData} element}

\item{j}{scalar character or numeric; 
name or index of coordinate space.}

\item{t}{transformation data.}

\item{...}{option arguments passed to and from other methods.}
}
\description{
Transformations
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

# image
q <- (p <- plotSpatialData()) + plotImage(x)
y <- x; image(y) <- scale(image(y), 1, c(1, 1, 1/3))
patchwork::wrap_plots(nrow=1, q, p + plotImage(y))
  
# point
point(x, "rotate") <- rotate(point(x), 20)
point(x, "scale") <- scale(point(x), c(1, 1.2))

plotSpatialData() +
  plotPoint(x) +
  plotPoint(x, "rotate", c="red") +
  plotPoint(x, "scale", c="blue")
  
# shape
shape(x, "rotate") <- rotate(shape(x), 5)
shape(x, "scale") <- scale(shape(x), c(1.2, 1))
shape(x, "translation") <- translation(shape(x), c(-5, 0))
plotSpatialData() +
  plotShape(x) +
  plotShape(x, "rotate", c="red") +
  plotShape(x, "scale", c="blue") +
  plotShape(x, "translation", c="gold") 
}
