% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\docType{data}
\name{trees_predictors}
\alias{trees_predictors}
\title{Predictor variable names for the dataset \code{trees}}
\format{
A character vector of length 50.
}
\usage{
data(trees_predictors)
}
\description{
Character vector of 50 predictor variable names from \link{trees}.
}
\examples{
data(trees_predictors)
trees_predictors
}
\seealso{
Other trees: 
\code{\link{trees}},
\code{\link{trees_extra}()},
\code{\link{trees_response}}
}
\concept{trees}
\keyword{datasets}
