% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\docType{data}
\name{trees_response}
\alias{trees_response}
\title{Response variable name for the dataset \code{trees}}
\format{
A character vector of length 1.
}
\usage{
data(trees_response)
}
\description{
Character vector of length 1 containing the name of the
response variable in \link{trees}.
}
\examples{
data(trees_response)
trees_response
}
\seealso{
Other trees: 
\code{\link{trees}},
\code{\link{trees_extra}()},
\code{\link{trees_predictors}}
}
\concept{trees}
\keyword{datasets}
