% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merfish_demo.R
\name{unzip_spd_demo}
\alias{unzip_spd_demo}
\title{unzip selected demo data to specified folder}
\usage{
unzip_spd_demo(
  zipname = "mibitof.zip",
  destination,
  cache = BiocFileCache::BiocFileCache(),
  source
)
}
\arguments{
\item{zipname}{character(1) should be name of zipped zarr archive found in Bioconductor OSN bucket}

\item{destination}{character(1) a path to a folder that must exist}

\item{cache}{defaults to BiocFileCache::BiocFileCache(), will serve as destination or source}

\item{source}{character(1) one of "biocOSN", "sandbox", "local"}
}
\value{
Returns path to base of unzipped archive
}
\description{
unzip selected demo data to specified folder
}
\examples{
tf <- tempfile()
dir.create(tf)
pa <- unzip_spd_demo(zipname="mibitof.zip", destination=tf, source="biocOSN")
dir(pa, full.names=TRUE)
}
