% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdio.R
\name{use_sdio}
\alias{use_sdio}
\title{use a reader}
\usage{
use_sdio(platform = "xenium")
}
\description{
use a reader
}
\examples{
use_sdio("xenium") # not basilisk-standards compliant -- leaks python reference into R
}
