% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{centroids}
\alias{extent}
\alias{centroids,ANY-method}
\alias{centroids,LabelArray-method}
\alias{centroids,ShapeFrame-method}
\alias{centroids,PointFrame-method}
\alias{extent,SpatialData-method}
\alias{extent,SpatialDataElement-method}
\title{Utilities}
\usage{
\S4method{centroids}{ANY}(x, ...)

\S4method{centroids}{LabelArray}(x, as = c("data.frame", "matrix"))

\S4method{centroids}{ShapeFrame}(x, as = c("data.frame", "matrix", "list"))

\S4method{centroids}{PointFrame}(x, as = c("data.frame", "list"))

\S4method{extent}{SpatialData}(x)

\S4method{extent}{SpatialDataElement}(x)
}
\arguments{
\item{x}{a \code{SpatialData} element (any but image).}

\item{...}{optional arguments passed to and from other methods.}

\item{as}{character string; how results should be returned.}
}
\value{
For \code{centroids}, a table (\code{data.frame} or \code{matrix})
of spatial coordinates (if \code{as="list"}, split by instance);
for extend, a length-2 numeric list of x- and y-ranges.
}
\description{
Utilities
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

centroids(label(x))
centroids(shape(x))
centroids(shape(x, 3), "list")

head(centroids(point(x)))
xy <- centroids(point(x), "list")
plot(xy$gene_a[, -3], col=a <- "red")
points(xy$gene_b[, -3], col=b <- "blue")
legend("topright", legend=names(xy), col=c(a, b), pch=21)

# object-wide
extent(x)

# element-wise
extent(label(x))
extent(point(x))
extent(shape(x))
}
