% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_predictors}
\alias{vi_predictors}
\title{Vector of all predictor names in \code{vi}}
\format{
Character vector of length 58.
}
\usage{
data(vi_predictors)
}
\description{
Vector of all predictor names in \code{vi}
}
\seealso{
Other vi: 
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_responses}}
}
\concept{vi}
\keyword{datasets}
