% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi_predictors}
\alias{vi_predictors}
\title{Vector of all predictor names in \code{vi}}
\format{
Character vector of length 58.
}
\usage{
data(vi_predictors)
}
\description{
Vector of all predictor names in \code{vi}
}
\seealso{
Other example_data: 
\code{\link{vi}},
\code{\link{vi_predictors_categorical}},
\code{\link{vi_predictors_numeric}},
\code{\link{vi_responses}}
}
\concept{example_data}
\keyword{datasets}
