% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_predictors}
\alias{vi_predictors}
\title{Predictor variable names for the dataset \code{vi}}
\format{
A character vector of length 58.
}
\usage{
data(vi_predictors)
}
\description{
Character vector of 58 predictor variable names from \link{vi}.
}
\examples{
data(vi_predictors)
vi_predictors
}
\seealso{
Other vi:
\code{\link{vi}},
\code{\link[=vi_extra]{vi_extra()}},
\code{\link{vi_responses}},
\code{\link{vi_smol}}
}
\concept{vi}
\keyword{datasets}
