% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_responses}
\alias{vi_responses}
\title{Vector of response names in \code{vi}}
\format{
Character vector with response names.
}
\usage{
data(vi_responses)
}
\description{
Vector of response names in \code{vi}
}
\seealso{
Other vi: 
\code{\link{vi}},
\code{\link{vi_extra}()},
\code{\link{vi_predictors}}
}
\concept{vi}
\keyword{datasets}
