% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\docType{data}
\name{vi_responses}
\alias{vi_responses}
\title{Response variable names for the dataset \code{vi}}
\format{
A character vector of length 5.
}
\usage{
data(vi_responses)
}
\description{
Character vector containing the names of the 5 response variables in \link{vi}.
}
\examples{
data(vi_responses)
vi_responses
}
\seealso{
Other vi:
\code{\link{vi}},
\code{\link[=vi_extra]{vi_extra()}},
\code{\link{vi_predictors}},
\code{\link{vi_smol}}
}
\concept{vi}
\keyword{datasets}
