% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeImageArray.R
\name{writeImageArray}
\alias{writeImageArray}
\title{Write `ImageArray` to Zarr-array}
\usage{
writeImageArray(image, path, ...)
}
\arguments{
\item{image}{A "ImageArray" specifying the image
to be saved.}

\item{path}{A character string specifying
a path}

\item{...}{Further arguments to be passed to write_zarr_array.}
}
\value{
\code{NULL}
}
\description{
...
}
\examples{
path <- "/path/to/my/image.zarr"
path <- system.file(path, package = "SpatialData")
writeImageArray(image, path)

}
