% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeSpatialData}
\alias{writeSpatialData}
\alias{writeImage}
\alias{writeLabel}
\alias{writePoint}
\alias{writeShape}
\alias{writeTable}
\title{Writing `SpatialData`}
\usage{
writeSpatialData(x, path, replace = TRUE, version = "0.2", ...)

writePoint(x, name, path, replace = TRUE, version = "0.2")

writeShape(x, name, path, replace = TRUE, version = "0.3")

writeImage(x, name, path, replace = TRUE, version = "0.3")

writeLabel(x, name, path, replace = TRUE, version = "0.3")
}
\arguments{
\item{x}{For \code{writeSpatialData},
a \code{SpatialData}
For \code{writeImage/Label/Point/Shape/Table}, 
a \code{ImageArray},\code{LabelArray}, 
\code{PointFrame}, \code{ShapeFrame}}

\item{path}{path to zarr store.}

\item{replace}{if TRUE, existing elements with the same name will be
replaced with the given element}

\item{version}{SpatialData version, 0.1 (zarr v2) or 0.2 (zarr v3)}

\item{...}{option arguments passed to and from other methods.}
}
\value{
\itemize{
\item{For \code{writeSpatialData}, a \code{SpatialData}.},
\item{For element writers, a \code{ImageArray}, \code{LabelArray}, 
\code{PointFrame}, \code{ShapeFrame}, or \code{SingleCellExperiment}.}}
}
\description{
Writing `SpatialData`
}
