% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeSpatialData}
\alias{writeSpatialData}
\alias{writeImage}
\alias{writeLabel}
\alias{writePoint}
\alias{writeShape}
\alias{writeTable}
\title{Writing `SpatialData`}
\usage{
writeSpatialData(x, name, path, replace = TRUE, version = "v2", ...)

writePoint(x, name, path, replace = TRUE, version = "v2")

writeShape(x, name, path, replace = TRUE, version = "v2")

writeImage(x, name, path, replace = TRUE, version = "v2")

writeLabel(x, name, path, replace = TRUE, version = "v2")
}
\arguments{
\item{x}{For \code{writeSpatialData},
a \code{SpatialData}
For \code{writeImage/Label/Point/Shape/Table}, 
a \code{ImageArray},\code{LabelArray}, 
\code{PointFrame}, \code{ShapeFrame}}

\item{name}{For \code{writeSpatialData},
name of the zarr store
For \code{writeImage/Label/Point/Shape/Table}, 
name of spatial element to write in the zarr store}

\item{path}{path to zarr store.}

\item{replace}{if TRUE, existing elements with the same name will be
replaced with the given element}

\item{version}{zarr version, v2 or v3 (only v2 is supported now)}

\item{...}{option arguments passed to and from other methods.}
}
\value{
\itemize{
\item{For \code{writeSpatialData}, a \code{SpatialData}.},
\item{For element writers, a \code{ImageArray}, \code{LabelArray}, 
\code{PointFrame}, \code{ShapeFrame}, or \code{SingleCellExperiment}.}}
}
\description{
Writing `SpatialData`
}
