% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zattrs.R
\name{Zattrs}
\alias{Zattrs}
\alias{$,Zattrs-method}
\alias{axes,Zattrs-method}
\alias{axes,SpatialDataElement-method}
\title{The `Zattrs` class}
\usage{
Zattrs(x = list())

\S4method{$}{Zattrs}(x, name)

\S4method{axes}{Zattrs}(x, ...)

\S4method{axes}{SpatialDataElement}(x, ...)
}
\arguments{
\item{x}{list extracted from a OME-NGFF compliant .zattrs file.}
}
\value{
\code{Zattrs}
}
\description{
The `Zattrs` class
}
\examples{
Zattrs()

}
