% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zattrs.R
\name{Zattrs}
\alias{Zattrs}
\title{The `Zattrs` class}
\usage{
Zattrs(x = list())
}
\arguments{
\item{x}{list extracted from a OME-NGFF compliant .zattrs file.}
}
\value{
\code{Zattrs}
}
\description{
The `Zattrs` class
}
\examples{
Zattrs()

}
