% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zattrs.R
\name{Zattrs}
\alias{Zattrs}
\alias{ct_data}
\alias{ct_axes}
\alias{ct_type}
\alias{ct_name}
\alias{ct_args}
\title{`SpatialData` .zattrs}
\usage{
ct_data(x, ...)

ct_type(x, ...)

ct_name(x, ...)

ct_axes(x, ...)

ct_args(x, i, ...)
}
\arguments{
\item{x}{\code{Zattrs}}

\item{...}{ignored.}

\item{i}{scalar integer or string 
specifying a coordinate transformation.}
}
\value{
\item{ct_axes}{character vector of axes names and types.}
\item{ct_data}{full coordinate transformation (CT) data.}
\item{ct_args}{CT arguments (scalar, vector, or matrix).}
\item{ct_type/name}{character vector for CT types/names.}
}
\description{
Handling of coordinate transformations.
}
\examples{
pa <- file.path("extdata", "blobs.zarr")
pa <- system.file(pa, package="SpatialData")
sd <- readSpatialData(pa)
za <- zattrs(sd@labels[[1]])

ct_axes(za)
ct_name(za)
ct_args(za, "scale")
ct_args(za, "affine")
}
