% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zattrs.R
\name{Zattrs}
\alias{Zattrs}
\alias{$,Zattrs-method}
\title{The `Zattrs` class}
\usage{
Zattrs(x = list())

\S4method{$}{Zattrs}(x, name)
}
\arguments{
\item{x}{list extracted from a OME-NGFF compliant .zattrs file.}

\item{name}{character string for extraction (see ?base::`$`).}
}
\value{
\code{Zattrs}
}
\description{
The `Zattrs` class
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- readSpatialData(x, tables=FALSE)

z <- meta(label(x))
axes(z) 
CTdata(z)
CTname(z)
CTtype(z)

}
